### 2020-01-29 #366 @jonkeane
dittodb
Submitting Author: Jonathan Keane (@jonkeane)
Repository:   http://github.com/jonkeane/dittodb

---

```
Package: dittodb
Title: A Test Environment for DB Requests
Version: 0.0.0.9000
    person(given = "Jonathan",
           family = "Keane",
           email = "jkeane@gmail.com",
    person(given = "Mauricio",
          family = "Vargas",
          role = c("aut"),
          email = "mvargas@dcc.uchile.cl",
  )
Description: Testing and documenting code that communicates with remote
  databases can be painful. Although the interaction with R is usually relatively
  simple (e.g. data(frames) passed to and from a database), because they rely on
  a separate service and the data there, testing them can be difficult to set up,
  unsustainable in a CI environment, or impossible without replicating an entire
  production cluster. This package addresses that by allowing you to make
  recordings from your database interactions and then play them back while
  testing (or in other contexts) all without needing to spin up or have access
  to the database your code would typically connect to.
License: Apache License (>= 2.0)
    R (>= 3.3.0),
    DBI
    digest,
    glue,
    callr,
    covr,
    dbplyr,
    nycflights13,
    odbc,
    RMariaDB,
    RPostgres,
    RPostgreSQL,
    RSQLite,
    withr
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

`dittodb ` provides a framework for testing R functions that interact with databases via the DBI interface. It allows one to use mocked database responses (and record those responses) to make it easier (or possible) to develop tests for functions that interact with databases.

Developers of packages and code that interact with databases

Not that I'm aware of.

Is this package a good fit for ROpenSci and the review process?
