### 2020-01-29 e5de0aa
Vignette infrastructure, add a testing section to readme
