### 2020-02-09 #366 @maurolepore
@jonkeane,

Thanks a lot for your submission! I'm excited to be the editor.

Before I do editor checks, I have two questions to follow up on some information that @maelle passed me:

1. Do you plan to rename the package?
2. Do you plan to add a vignette?
