### 2020-02-10 #366 @jonkeane
Hello! Thanks for the message. Yes, I'm planning to change the name (though not 100% determine on to what just yet) and I will be adding one more vignette that gives a getting started / overview of the package. I will let you know when both of those are done (I'm hoping this week or so)
