### 2020-02-11 53e4c43
Shuffling while writting vignettes
