+++ vignettes/travelling.Rmd
@@ -13,7 +13,7 @@ vignette: >
 library(dbtest)
 
 # set the mockPaths for this vignette
-.mockPaths("travelling")
+.db_mock_paths("travelling")
 
 knitr::opts_chunk$set(eval = TRUE, message = FALSE, warning = FALSE)
 ```
@@ -83,13 +83,13 @@ without a connection to the database.
 
 While we are able to connect to the database (i.e. when we are at the office) we
 can save the results returned by queries with code like the following (by
-calling `start_capturing()` before the connection and the code that executes the
-queries and then `stop_capturing()` at the end):
+calling `start_db_capturing()` before the connection and the code that executes the
+queries and then `stop_db_capturing()` at the end):
 
 ```{r, eval=FALSE}
 library(dbtest)
 
-start_capturing()
+start_db_capturing()
 
 con_psql <- RPostgreSQL::dbConnect(
     drv = DBI::dbDriver("PostgreSQL"),
@@ -110,7 +110,7 @@ flights_100
 
 dbDisconnect(con_psql)
 
-stop_capturing()
+stop_db_capturing()
 ```
 
 ```{r cooking show trick, echo=FALSE}

[…]
