### 2020-02-11 15e40ca
Remembering rmarkdown chunk options
