+++ vignettes/getting-started.Rmd
@@ -35,6 +35,13 @@ con <- dbConnect(
 )
 
 nycflights13_sql(con, schema = "nycflights13")
+
+# record interactions
+
+start_db_capturing()
+out <- mean_delays("day")
+out <- mean_delays("month")
+stop_db_capturing()
 ```
 
 ```{r mean_delays}
@@ -60,14 +67,14 @@ mean_delays <- function(group_col) {
 }
 ```
 
-If we give it the column `"day"`, we get the following dataframe:
+If we give it the column `"month"`, we get the following dataframe:
 
-```{r day, eval = FALSE}
-mean_delays("day")
+```{r month, eval = FALSE}
+mean_delays("month")
 ```
 
 ```{r cooking_show}
-with_mock_db(mean_delays("day"))
+with_mock_db(mean_delays("month"))
 ```
 
 Great, now that we have our function we want to test it to make sure it is operating as expected. Normally, we could write something like:
@@ -77,9 +84,9 @@ library(testthat)
 library(dbtest)
 
 test_that("mean_delays()", {
-  out <- mean_delays("day")
-  expect_named(out, c("day", "mean_delay"))
-  expect_equal(dim(out), c(31, 2))
+  out <- mean_delays("month")
+  expect_named(out, c("month", "mean_delay"))
+  expect_equal(dim(out), c(12, 2))
 })
 ```
 
@@ -91,11 +98,11 @@ We can record snapshots of the database interactions with the commands `start_db
 
 ```{r recording, eval = FALSE}
 start_db_capturing()
-out <- mean_delays("day")
+out <- mean_delays("month")
 stop_db_capturing()
 ```
 
-This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-16d120.R` which is the snapshot for this snapshot. This `SELECT-...` file contains the data that was received from the database for use in tests.
+This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-e53189.R` which is the snapshot for this snapshot. This `SELECT-...` file contains the data that was received from the database for use in tests.
 
 ## `with_mock_db()`
 
@@ -103,10 +110,9 @@ Now that we have a snapshot, we can use that snapshot by wrapping our call that 
 
 ```{r with_mock_1}
 with_mock_db(
-  mean_delays("day")
+  mean_delays("month")
 )
 ```
-
 
 So, now we can write our tests like:
 
@@ -116,6 +122,28 @@ library(dbtest)
 
 with_mock_db(
   test_that("mean_delays()", {
+    out <- mean_delays("month")
+    expect_named(out, c("month", "mean_delay"))
+    expect_equal(dim(out), c(12, 2))
+  })
+)
+```
+
+When placed inside of `with_mock_db(...)` a call to `mean_delays("month")` will return what we saved as our snapshot _as if it had actually connected to the database_ without needing the database to be installed, reachable, or operational.
+
+If we wanted to test that a day-based aggregation works, we can, although we will have to make a new snapshot. First we would run the following interactively:
+
+```{r recording_days, eval = FALSE}
+start_db_capturing()
+out <- mean_delays("day")
+stop_db_capturing()
+```
+
+This will create a new file (`SELECT-16d120.R`) which contains the response when aggregating by day. dbtest saves each database interaction with a hash of the query that is sent, so that a number of different responses from a database can be saved and the correct one will be used when called inside of `with_mock_db(...)`. So now, we could write our new test with:
+
+```{r tests_day_2}
+with_mock_db(
+  test_that("mean_delays()", {
     out <- mean_delays("day")
[…]
