### 2020-02-13 3ebe9d6
Finishing up first draft of vignette
