+++ vignettes/getting-started.Rmd
@@ -1,5 +1,7 @@
 ---
 title: "Getting Started with `dbtest`"
+author: "Jonathan Keane"
+date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{getting-started}
@@ -19,9 +21,9 @@ knitr::opts_chunk$set(
 .db_mock_paths("getting-started")
 ```
 
-`dbtest` is designed to make it easy and fun to test functions that interact with a database. It works be looking for mock responses for each query you send while you run your tests and will seamlessly pretend that those mocks were provided by the database connection without needing a connection at all.
+`dbtest` is designed to make it easy and fun to test functions that interact with a database. It works by looking for mock responses for each query you send while you run your tests and will seamlessly pretend that those mocks were provided by the database connection without needing a connection at all.
 
-To get started, imagine that we are working on a package that queries a database that consists of the [nycflights13 data][nycflights13_description]. We have the following function which takes a column to aggregate by and returns a dataframe with that column and the mean delay for groups based on the values in the column name given.
+To get started, imagine that we are working on a package that queries a database that consists of the [nycflights13 data](nycflights13_description). We have the following function which takes a column to aggregate by and returns a dataframe with that column and the mean delay for groups based on the values in the column name given.
 
 ```{r setup db, include = FALSE, eval = FALSE}
 # setup the DB used in the rest of the vignette
@@ -102,7 +104,7 @@ out <- mean_delays("month")
 stop_db_capturing()
 ```
 
-This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-e53189.R` which is the snapshot for this snapshot. This `SELECT-...` file contains the data that was received from the database for use in tests.
+This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-e53189.R` which is the snapshot for this example. This `SELECT-*` file contains the data that was received from the database for use in tests.
 
 ## `with_mock_db()`
 
@@ -155,11 +157,11 @@ with_mock_db(
 
 There are a few things to be careful about when using dbtest. 
 
-_When to call `dbConnect()`_
+### When to call `dbConnect()`
 
 Always call `dbConnect()` inside of `with_mock_db(...)`. You can make as many calls as you want to the mock database inside of a `with_mock_db(...)`, but you should always make sure that you connect to the database inside of and not outside of `with_mock_db(...)`. This is because when you "connect" to the mock database, a few variables are set that tell dbtest where to look for mocks. It's less important (though still a good idea) to call `dbDisconnect()` inside of `with_mock_db()`. This is also true when recording snapshots with `start_db_recording()`, you should start the recording and then call `dbConnect()`.
 
-_Query size_
+### Query size
 
 Recording snapshots saves the whole query to disk in a relatively inefficient way (from a data storage perspective), so be careful with what you save. And you'll want to not save extremely large results if at all possible. This is also a best-practice for writing tests anyway: you should have mocks that are as minimal as possible to test the functionality you need to. Minimal mocks make it easier to change things that aren't relevant to the test (you don't have to change the way data is represented if it's not important to what you're testing) and it makes your tests run faster.
 
@@ -178,11 +180,16 @@ You can add a new path by calling `.db_mock_paths("new/path/here")` which will a
 
 ### Redacting
 
[…]
