### 2020-02-13 e13037b
Add lintr
