### 2020-02-13 6e1d410
lintr update
