+++ R/redact.R
@@ -47,7 +47,7 @@ standard_redactors <- function(data, columns) {
 #'
 #' @return data, with the columns specified in `columns` duly redacted
 #' @export
-redact_columns <- function(data, columns, ignore.case = TRUE, ...) {
+redact_columns <- function(data, columns, ignore.case = TRUE, ...) { # nolint
   columns <- lapply(
     columns,
     grep,
