+++ R/mock-db.R
@@ -1,14 +1,0 @@
-#' Run the DBI queries inside against a mocked db
-#'
-#' @param expr the expression to execute
-#'
-#' @return nothing
-#' @importFrom testthat with_mock
-#' @export
-with_mock_db <- function(expr) {
-  with_mock(
-    dbConnect = function(...) dbMockConnect(...),
-    .env = "DBI",
-    expr
-  )
-}
