+++ .travis.yml
@@ -1,46 +1,0 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
-language: R
-cache: packages
-
-# linux builds should be able to use the docker set up, but have been
-# segfaulting on connection with either MariaDB or RPostgres
-before_install:
-  - if [ $TRAVIS_OS_NAME = "linux" ]; then sudo /etc/init.d/postgresql stop || true; bash db-setup/postgres-docker.sh; fi
-  - if [ $TRAVIS_OS_NAME = "osx" ]; then bash db-setup/postgres-brew.sh; fi
-  - if [ $TRAVIS_OS_NAME = "osx" ]; then bash db-setup/mariadb-brew.sh; fi
-  - sleep 10
-  - psql -h 127.0.0.1 -U postgres -f db-setup/postgres-reset.sql
-  - psql -h 127.0.0.1 -U postgres nycflights -f db-setup/postgres-nycflights.sql
-  - mysql -u root < db-setup/mariadb-reset.sql
-  - mysql -u root -e "source db-setup/mariadb-nycflights.sql"
-
-services:
-  - mysql
-
-addons:
-  apt:
-    packages:
-      # this is to make the odbc/psql packages work
-      - unixodbc-dev
-      - odbc-postgresql
-      - libpq-dev
-
-before_script:
-  - if [ $TRAVIS_OS_NAME = "linux" ]; then sudo apt install unixodbc-dev odbc-postgresql; fi
-  # we have to install from source, or else we get `result_fetch(res@ptr, n = n)`
-  - R -q -e 'install.packages(c("RMariaDB","odbc","RPostgreSQL","RPostgres"), type = "source");'
-
-r:
-  - oldrel
-  - release
-  - devel
-os:
-  - linux
-  - osx
-matrix:
-  exclude:
-  - os: osx
-    r: oldrel
-  - os: osx
-    r: devel
