+++ tests/testthat/helpers.R
@@ -1,11 +1,11 @@
 skip_env <- function(which = c("RPostgres", "RMariaDB")) {
   if ("RPostgres" %in% which &&
-      tolower(Sys.getenv("DBTEST_DISABLE_PG")) == "true") {
+      tolower(Sys.getenv("dittodb_DISABLE_PG")) == "true") {
     skip("Skipping tests that need functioning RPostgres.")
   }
 
   if ("RMariaDB" %in% which &&
-      tolower(Sys.getenv("DBTEST_DISABLE_MARIA")) == "true") {
+      tolower(Sys.getenv("dittodb_DISABLE_MARIA")) == "true") {
     skip("Skipping tests that need functioning RMariaDB.")
   }
 
