+++ README.md
@@ -1,15 +1,15 @@
-# dbtest
+# dittodb
 <!-- badges: start -->
-[![R build status](https://github.com/jonkeane/dbtest/workflows/R-CMD-check/badge.svg)](https://github.com/jonkeane/dbtest/actions?workflow=R-CMD-check)
-[![Codecov test coverage](https://codecov.io/gh/jonkeane/dbtest/branch/master/graph/badge.svg)](https://codecov.io/gh/jonkeane/dbtest?branch=master)
+[![R build status](https://github.com/jonkeane/dittodb/workflows/R-CMD-check/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=R-CMD-check)
+[![Codecov test coverage](https://codecov.io/gh/jonkeane/dittodb/branch/master/graph/badge.svg)](https://codecov.io/gh/jonkeane/dittodb?branch=master)
 [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![CRAN status](https://www.r-pkg.org/badges/version/dbtest)](https://CRAN.R-project.org/package=dbtest)
+[![CRAN status](https://www.r-pkg.org/badges/version/dittodb)](https://CRAN.R-project.org/package=dittodb)
 <!-- badges: end -->
 
 
-dbtest is a package that makes testing against databases easy. When writing code that relies on interactions with databases, testing has been difficult without recreating test databases in your CI environment, or resorting to using SQLite databases instead of the database engines you have in production. Both have their downsides: recreating database infrastructure is slow, error prone, and hard to iterate with. Using SQLite works well, right up until you use a feature (like [a full outer join](https://www.sqlite.org/omitted.html)) or has [quirks](https://www.sqlite.org/quirks.html) that might differ from your production database. dbtest solves this by recording database interactions, saving them as mocks, and then replaying them seamlessly during testing. This means that if you can get a query from your database, you can record the response and reliably reproduce that response in tests.
+dittodb is a package that makes testing against databases easy. When writing code that relies on interactions with databases, testing has been difficult without recreating test databases in your CI environment, or resorting to using SQLite databases instead of the database engines you have in production. Both have their downsides: recreating database infrastructure is slow, error prone, and hard to iterate with. Using SQLite works well, right up until you use a feature (like [a full outer join](https://www.sqlite.org/omitted.html)) or has [quirks](https://www.sqlite.org/quirks.html) that might differ from your production database. dittodb solves this by recording database interactions, saving them as mocks, and then replaying them seamlessly during testing. This means that if you can get a query from your database, you can record the response and reliably reproduce that response in tests.
 
-dbtest is heavily inspired by [httptest](https://CRAN.R-project.org/package=httptest), if you've used httptest before, you'll find many of the interactions similar.
+dittodb is heavily inspired by [httptest](https://CRAN.R-project.org/package=httptest), if you've used httptest before, you'll find many of the interactions similar.
 
 ## A quick example
 Say we have a Postgres database with some [`nycflights`](https://CRAN.R-project.org/package=nycflights13) data in it and we are writing functions that query this data that we want to test.
@@ -109,20 +109,20 @@ with_mock_db({
 ```
 
 ## Installation
-Currently, dbtest is not on CRAN. You can install from source, or use `devtools`:
+Currently, dittodb is not on CRAN. You can install from source, or use `devtools`:
 
 ```r
-devtools::install_github("jonkeane/dbtest")
+devtools::install_github("jonkeane/dittodb")
 ```
 
 ## Development
[…]
