+++ man/dot-dittodb_env.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/recording.R
 \docType{data}
-\name{.dbtest_env}
-\alias{.dbtest_env}
-\title{an environment for dbtest storing state}
+\name{.dittodb_env}
+\alias{.dittodb_env}
+\title{an environment for dittodb storing state}
 \format{An object of class \code{environment} of length 0.}
 \usage{
-.dbtest_env
+.dittodb_env
 }
 \description{
-an environment for dbtest storing state
+an environment for dittodb storing state
 }
 \keyword{internal}
