+++ NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
 export(.db_mock_paths)
-export(.dbtest_env)
+export(.dittodb_env)
 export(.mockPaths)
 export(db_path_sanitize)
-export(dbtest_debug_level)
+export(dittodb_debug_level)
 export(get_dbname)
 export(get_redactor)
 export(get_type)
