### 2020-02-14 2906389
Remove direct link to httptest's .mockPaths() to facilitate cross reference checkability
