### 2020-02-14 #366 @jonkeane
Ok, I've renamed the package (and changed the github repo name, pkgdown site, etc.) the new name is `dittodb` and it's at https://github.com/jonkeane/dittodb

I've finished up a relatively comprehensive getting started vignette https://dittodb.jonkeane.com/articles/getting-started.html

So this should be good to go.
