+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: dittodb
 Title: A Test Environment for DB Requests
-Version: 0.1.0 
+Version: 0.1.0
 Authors@R: c(
     person(given = "Jonathan",
            family = "Keane",
