### 2020-02-14 c57bfb0
example with skimr
