### 2020-02-14 a1d3f16
removes nycflights vignette
