+++ vignettes/nycflights.Rmd
@@ -1,86 +1,0 @@
----
-title: "nycflights13 data"
-author: "Mauricio Vargas and Jonathan Keane"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{nycflights13_description}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include=FALSE}
-library(dittodb)
-
-# set the mockPaths for this vignette
-.db_mock_paths("nycflights13")
-
-knitr::opts_chunk$set(eval = TRUE, message = FALSE, warning = FALSE)
-```
-
-`dittodb` uses the `nycflights13` dataset (from the `nycflights13` package) for 
-testing and example purposes.
-
-# Exploring `nycflights13`
-
-The `nycflights13` dataset contains airline on-time data for all flights 
-departing NYC in 2013. It also includes useful 'metadata' on airlines, airports,
-weather, and planes.
-    
-Have a look to the database schema:
-
-![`nycflights13` relational diagram.](relational-nycflights.svg)
-
-Let's create an SQLite connection, the database will be created on the fly:
-```{r}
-con <- nycflights13_sqlite()
-```
-
-Now let's explore the data by using `dplyr` and `skimr`:
-```{r}
-library(dplyr)
-library(dbplyr)
-library(skimr)
-
-# define a function to avoid repeating myself
-db_descriptive_stats <- function(table) {
-  tbl(con, table) %>% 
-    skim() %>% 
-    mutate_if(is.numeric, function(x) { round(x, 2) })
-}
-
-tables <- c("airports", "flights",  "weather", "planes", "airlines")
-
-explore <- lapply(tables, db_descriptive_stats)
-names(explore) <- tables
-```
-
-## Airports
-
-```{r}
-explore$airports
-```
-
-## Flights
-
-```{r}
-explore$flights
-```
-
-## Weather
-
-```{r}
-explore$weather
-```
-
-## Planes
-
-```{r}
-explore$planes
-```
-
-## Airlines
-
-```{r}
-explore$airlines
-```
