+++ DESCRIPTION
@@ -40,7 +40,6 @@ Suggests:
     covr,
     dplyr,
     dbplyr,
-    skimr,
     knitr,
     nycflights13,
     odbc,
