+++ .github/workflows/R-CMD-check.yaml
@@ -79,7 +79,7 @@ jobs:
           sudo sed -i 's/Driver=psqlodbca.so/Driver=\/usr\/lib\/x86_64-linux-gnu\/odbc\/psqlodbca.so/g' /etc/odbcinst.ini
           # mysql needs time to boot, try it every 5 seconds for 1 minute
           n=0
-          until [ $n -ge 12 ]
+          until [ $n -ge 25 ]
           do
             mysql -h 127.0.0.1 -u root -e "SHOW DATABASES" && break
             sleep 5
