### 2020-02-15 #366 @maurolepore
### Editor checks:

- [ ] **Fit**: The package meets[^1] criteria[^2] for [fit](policies.md#fit) and [overlap](policies.md#fit)[^3]

---

#### Editor comments

[^1]: @maelle, I see a conflict of interest. [Mauricio Vargas authors dittodb](https://github.com/jonkeane/dittodb/blob/d8ab0725f783cb74a55517ee1cf1fa9c566958ef/DESCRIPTION#L10) and [co-authors with me the datos package](https://CRAN.R-project.org/package=datos ), first published on 2019-05-31 ([archive](https://cran.r-project.org/src/contrib/Archive/datos/)). The [policies](https://devguide.ropensci.org/policies.html#coi) state: _"The potential editor or reviewer has a conflict of interest if: Within in the past three years, the potential reviewer/editor has been a collaborator or has had any other professional relationship with any person on the package who has a major role"_. While I'm here, I'll add continue with the editor's checks but please let me know how to proceed.

[^2]:

- 2.1. "database access" is listed under "Scope" in the [first comment of this issue](https://github.com/ropensci/software-review/issues/366#issue-557163606) but [I see no such category in the polices document](https://devguide.ropensci.org/policies.html#package-categories).

- 2.2. The categories "data validation and testing" and "workflow automation" seem related to the goal of dittodb.

- 2.3. The policy document states _"Here are some types of packages we are unlikely to accept: ... General workflow or package development support packages"_, and @jonkeane states that the target audience of dittodb is: _"Developers of packages and code that interact with databases"_. @jonkeane, could you please clarify how scientists can benefit from dittodb?

[^3] README.md states: _"dittodb is heavily inspired by [httptest](https://CRAN.R-project.org/package=httptest)"_. @jonkeane, could you please consider this and see if you can add expand you answer to _"Are there other R packages that accomplish the same thing? If so, how does yours differ or meet [our criteria for best-in-category](https://devguide.ropensci.org/policies.html#overlap)?"_ (see [first comment of this issue](https://github.com/ropensci/software-review/issues/366#issue-557163606)).

Regarding what's listed under [Upon submission](https://devguide.ropensci.org/editorguide.html#editorchecklist), I see ` goodpractice::gp()` gives these notes (_"[most exceptions will need to be justified by the author in the particular context of their package](https://devguide.ropensci.org/editorguide.html#editorchecklist)"_):

```
It is good practice to

    R/connection.R:56:NA
    R/connection.R:57:NA
    R/connection.R:59:NA
    R/connection.R:60:NA
    R/connection.R:62:NA
    ... and 80 more lines

  ✖ not use "Depends" in DESCRIPTION, as it can cause name
    clashes, and poor interaction with other packages. Use "Imports"
    instead.

    R/driver-specific-connections.R:63:1
    R/recording.R:122:1
    R/result.R:98:1
    R/result.R:114:1
    tests/testthat/test-connection.R:7:1

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check WARNING: LaTeX errors when creating
    PDF version. This typically indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running with no
    redirection of stdout/stderr. Hmm ... looks like a package Error in
    texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, :
    pdflatex is not available Error in texi2dvi(file = file, pdf =
    TRUE, clean = clean, quiet = quiet, : pdflatex is not available
    Error in running tools::texi2pdf() You may want to clean up by 'rm
    -Rf /tmp/Rtmp2o3RWi/Rd2pdf34286ec86563'
```

---

Reviewers: Pending until @maelle tells me how to proceed about the conflict of interest I explained above.
Due date:
