+++ R/recording.R
@@ -119,7 +119,10 @@ start_db_capturing <- function(path, redact_columns = NULL) {
   quietly(trace_dbi(
     "dbConnect",
     exit = quote({
-      .dittodb_env$db_path <- file.path(.db_mock_paths()[1], get_dbname(list(...)))
+      .dittodb_env$db_path <- file.path(
+        .db_mock_paths()[1],
+        get_dbname(list(...))
+      )
       dir.create(.dittodb_env$db_path, showWarnings = FALSE, recursive = TRUE)
     })
   ))
