### 2020-02-15 #366 @pachamaltese
@maelle At the risk of sounding like Mr. Peanutbutter, Mauro has collaborated with us at the R4DS spanish translation and I only find him to be a very nice person, he even understood when I couldn't take the plane to Buenos Aires for the RUG meetup there, but that can be a conflict of interest.
