### 2020-02-15 #366 @jonkeane
Hello, thank you for the detailed notes. Here are my responses to the above questions:

[^2]
2.1 and 2.2 I'm not sure that these are directed at me. If they are, let me know and I will address them.

2.3 This benefits scientist by providing them with a tool that makes it easy(er) to test code they write that interacts with databases. Connecting with and interacting with databases is core to many workflows and it is critical that this code work correctly and `dittodb` provides a framework and functionality to write tests for that code that is otherwise quite difficult to write on its own.

[^3]
The `httptest` package is limited to testing http-based API interactions. Though `dittodb` takes a similar approach in the way that it works, it's solving a different problem.  I've looked through https://ropensci.org/packages/ with the keywords `database`, `db`, and `test` and don't see any packages that provide functionality to facilitate writing and running tests against databases like `dittodb`. Further, I've searched both CRAN and google/github for packages that do this already and have not found any. The few database and test related packages I have found address different needs:
* `DBItest` (https://github.com/r-dbi/DBItest) is designed and intended to test new DBI-compliant backends
* `dbtest` (https://github.com/rstudio/dbtest) is designed and intended to test `dbplyr` translations for backends

# goodpractice::gp() comments

`write unit tests for all functions, and all package code in general. 79% of code lines are covered by test cases.`
`dittodb` has 100% test coverage when run on CI (and when run with specific configurations locally) Because of the complexity of testing database connections (and specifically spinning up the test DBs to test against locally), full test running is disabled by default locally. We use GitHub Actions for CI, and there the CI process installs the correct databases and runs all of the tests.
https://codecov.io/gh/jonkeane/dittodb?branch=master

`not use "Depends" in DESCRIPTION, as it can cause name clashes, and poor interaction with other packages. Use "Imports"  instead.`
For most of the dependencies `dittodb` uses `Imports`, but for `DBI` we chose to use `Depends`. This is because `dittodb` is designed to be use in conjunction with and extend `DBI`. This follows what packages like https://github.com/r-dbi/RPostgres do as well as the advice given at the end of the section in http://r-pkgs.had.co.nz/namespace.html#search-path

`avoid long code lines, it is bad for readability. Also, many people prefer editor windows that are about 80 characters wide. Try make your lines shorter than 80 characters`
I've fixed two of these, the remaining 3 lines are comments with long URLs referencing relevant issues. I can add a line break in the URL, though that makes accessing the links more difficult.

`not import packages as a whole, as this can cause name clashes between the imported packages. Instead, import only the specific functions you need.`
This appears to be a consequence of using `Depends` for `DBI` which was addressed above.

`fix this R CMD check WARNING: LaTeX errors when creating ...` and
`fix this R CMD check ERROR: Re-running with no ...`
Both of these look like they are problems with the computer that `good practice::gp()` was running on not having an installation of tex that R could find (the error being: `pdflatex is not available`). I can build the package and PDF manually locally (as well as on CI) and `goodpractice::gp()` runs without these errors locally. Would you mind trying to install Yihui's `tinytex` (https://yihui.org/tinytex/) and see if that resolves those two errors?

```
install.packages('tinytex')
tinytex::install_tinytex()
```
