### 2020-02-18 5a2e803
A bit of clean up
