+++ .github/workflows/check-windows.yaml
@@ -71,14 +71,10 @@ jobs:
       - name: Install R dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'httptest'), type = 'source')" -e "remotes::install_deps(dependencies = TRUE);"
 
-      - name: Build R DB Dependencies on cache miss
-        if: steps.rlibcache.outputs.cache-hit != 'true' && contains(runner.os, 'windows') == false
-        run: Rscript -e "install.packages(c('RMariaDB', 'odbc', 'RPostgreSQL', 'RPostgres'), type = 'source');"
-
-      - name: Build RPostgres on windows
+      - name: Build RPostgres on windows on cache miss
         # RMariaDB and RPostgres interact in funny ways on windows:
         # https://github.com/r-dbi/RMariaDB/issues/119
-        if: steps.rlibcache.outputs.cache-hit != 'true' && contains(runner.os, 'windows')
+        if: steps.rlibcache.outputs.cache-hit != 'true'
         run: Rscript -e "remotes::install_github('r-dbi/RPostgres');"
 
       - name: Setup databases
