### 2020-02-19 96bfcc0
Move db population to script
