+++ .github/workflows/check-linux.yaml
@@ -18,7 +18,6 @@ jobs:
         - { os: ubuntu-18.04, r: 'release', args: "--no-manual" }
 
     env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       CRAN: ${{ matrix.config.cran }}
       DBROOTPASS: r2N5y7V*
 
