+++ .github/workflows/check-windows.yaml
@@ -15,7 +15,6 @@ jobs:
         - { os: windows-latest, r: 'release', args: "--no-manual" }
 
     env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       CRAN: ${{ matrix.config.cran }}
       DBROOTPASS: r2N5y7V*
 
