### 2020-02-19 38772ec
Add R_REMOTES back to mac
