### 2020-02-19 #366 @annakrystalli
Dear @maurolepore ,

Thank you for bringing to our attention a potential conflict of interest. We've discussed and concluded is it best to address the COI so if you're happy the editors checks are complete, I'll take over from here and start looking for reviewer. We really appreciate your offer to guest edit so we'd be really happy if you consider handling another submission when it comes up?

Hi  @jonkeane  👋, I'll be taking over from here as editor.
