+++ .github/workflows/check-linux-ubuntu.yaml
@@ -33,11 +33,6 @@ jobs:
       - uses: r-lib/actions/setup-tinytex@master
         if: contains(matrix.config.args, 'no-manual') == false
 
-      - name: Ensure that tex has all the correct packages
-        if: contains(matrix.config.args, 'no-manual') == false
-        run: |
-          tlmgr install gettitlestring
-
       - name: Cache R packages
         uses: actions/cache@v1
         id: rlibcache
