### 2020-02-20 c7db467
Remove object usage linter because it seems to be throwing errant errors?
