+++ db-setup/postgres-brew.sh
@@ -2,7 +2,8 @@ rm -rf /usr/local/var/postgres
 rm -rf /usr/local/bin/psql
 # for odbc tests, we need to also install the odbc drivers
 # we might also need to install unixodbc here if it's not already installed.
-brew install postgresql psqlodbc
+brew install postgresql
+brew install psqlodbc
 cat <<EOT >> /usr/local/etc/odbcinst.ini
 [PostgreSQL Unicode]
 Description     = PostgreSQL ODBC driver (Unicode 9.2)
