### 2020-02-20 jonkeane/dittodb#61 @jonkeane
Best approach for custom serializers?
`dput`s of objects were chosen as the first serialization for a few reasons:

1. they are plain text so easily reviewable and understandable in git diffs
1. they serialize any sort of object
1. they can be used to reliably return a data.frame with specific column types

A few alternatives were not chosen:

*CSV*
While these are plain text, and arguably easier to read than `dput` output, they would need some sort of sidecar file to make sure they are parsed correctly into data.frames and they couldn't be used to serialize non-data.frame objects. (missing 2 and 3 above)

*RDS*
These can serialize anything (and reliably return data.frames but they don't satisfy (1) above since they are binary and not plain text.
