### 2020-02-21 cbd90a7
separate gh actions (#60)
Also add lighting to macOS release build

Co-authored-by: Jonathan Keane <jkeane@gmail.com>
