+++ README.md
@@ -1,6 +1,8 @@
 # dittodb
 <!-- badges: start -->
-[![R build status](https://github.com/jonkeane/dittodb/workflows/R-CMD-check/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=R-CMD-check)
+[![Windows](https://github.com/jonkeane/dittodb/workflows/check-windows/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-windows)
+[![Mac](https://github.com/jonkeane/dittodb/workflows/check-mac/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-mac)
+[![Linux](https://github.com/jonkeane/dittodb/workflows/check-linux-ubuntu/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-linux-ubuntu)
 [![Codecov test coverage](https://codecov.io/gh/jonkeane/dittodb/branch/master/graph/badge.svg)](https://codecov.io/gh/jonkeane/dittodb?branch=master)
 [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](https://www.r-pkg.org/badges/version/dittodb)](https://CRAN.R-project.org/package=dittodb)
