### 2020-03-24 #366 @etiennebr
Thanks @annakrystalli. I look forward to reviewing this package and despite that my agenda is scrambled, at this point, I think that I can review dittodb before April 18.
