### 2020-04-17 #366 @etiennebr
- [-] **Performance:** Any performance claims of the software been confirmed.

---

#### Checklist

I will start by commenting on the unchecked items above.

##### Installation instructions

The installation is straightforward, but I think the installation of databases,
a major piece of the package, could be improved to help on-board users. I first
setup a container to run the examples, then while going through the tests I saw
there were scripts already available that would just need to be documented for
the user.

##### Vignettes

The Vignettes are relevant, although slightly redundant. Since the package is
focused on tests, I would suggest that a vignette walking the user through the
full process of testing during package development would make adoption much
easier. I'll comment further on this later, but I could not figure a smooth
workflow, which I'm sure the authors have figured out.

I find the diversity of examples in the vignettes and the `README` very
relevant, but as I mentionned above, database setup could be improved to
accelerate the users. I would suggest to:

1. Provide a simple example using maybe a simple `sqlite` file in the readme;
2. Provide instructions to run a docker container in the vignettes.

##### Community guidelines

There is a code of conduct, but I couldn't find instructions for contributions.

##### Automated tests

I like the fact the the package targets 100% coverage. I know it doesn't mean
it's perfectly tested, but I find it easier to follow `nocov` tags in the code
(in this case there are none) and know that the rest is tested. However, when I
ran the tests locally, I only got 80% coverage. I see that some tests are
skipped locally even if I run both `db-setup/postgres-docker.sh` and
`db-setup/mariadb-docker.sh` as suggested (I believe it should be the
`*-reset.sh` version). Maybe I missed it, but I think it would be useful to tell
the contributors why some tests can't be run locally.

`dittodb` identified a real problem for me and I am not aware of any other
package that tries to solve that same problem. The `DESCRIPTION` and the
`README` clearly state the problem of running tests against a database and that
`dittodb`'s solution is to mock a database connection to simulate the database
response. This saves the setup of a database in CI, and I would add, saves
contributors to have to setup a local database to run integration tests if they
don't touch the db interface.

##### Workflow

As I mentioned earlier, I could understand the vignettes and the examples, but
I'm not sure I understand `dittodb` intented package development workflow. By
looking at the package's own tests, I think I could grapple some ideas, but I
still find the ergonomic of the package could be better (or explained). Maybe I
should just get more familiar with it, but I feel like I spent a larger amount
of time playing with it than most potential users would --out of pure
motivation, mind you.

Here's how I used `dittodb`:
- add a ` start` and `stop` line at the beginning and the end of the database tests
- run the tests manually
- remove the `start` and `stop` line
- wrap all the code in a `with_mock_db({})`
- repeat every time you add a test (and make sure contributors know the gymnastic)

Here are my pain points:
- Debugging is harder, because I can't step through the tests
- We need to change many lines of code in different places every time a test
changes, which makes it error prone
- The commit that adds `dittodb` change all the lines in a test file
(re-indent), which burries the changes in another layer (maybe that's just me)

What I would like is to:
- Make it possible to run tests line-by-line
- Set it up once and not think about it (especially other contributors)
- Have minimal changes in the code

I came up with a suggestion, but I have no idea of the feasability (although I
think it could be possible). Maybe `dittodb` could strictly modify the database
connection with a custom e.g. `dittodb_con` class and use normal dispatch to
catch results. So if the connection is possible, it connects to the database. If
the connection is not possible, it will use fixtures. That way if the dev has a
working database connection, it will run the tests against the database and
write fixtures, but if the connection is impossible, it will run against the
fixture. That way we can track the response by committing fixtures (which is
cool) and see if a change requires a change in fixtures (and it's transparent).

Maybe it involves to much changes or it simply is not how you see things, or
maybe you see difficulties I overlooked. Clearly, you put more thoughts into it
than me. But just to make it clear and let you judge it, here's how I think the
UX could be:
```r
con <- ditto_connect(  # to connect if possible, otherwise use the recorded session
  RMariaDB::MariaDB(),
  dbname = "nycflights",
  host = "127.0.0.1",
  username = "travis",
  password = "",
  ditto_file = "<optional file name>"
)

ditto_connect <- function(..., ditto_file, redacted) {
  con <- tryCatch(
    DBI::dbConnect(...),
    error = mock_db(..., ditto_file)  # would probably need something more robust
  )
  structure(con, c("ditto_con", class(con)))
}

# The rest is not trivial, but I believe a lot of the needed infrastructure of
# `dittodb` exists to catch when this connection fetches records and write
# to disc, or use the fixtures.

# Using the examples from the vignette
mean_delays(con, grou_col)
```

Before coming with this idea, my initial thought was to provide a similar syntax
to record and mock the database to reduce the number of lines to change in the
code when going from record to replay, which I think could a a good fall back.
This was my initial thought:
```r
record_mock_db({  # instead of start_db_capturing()
  out <- mean_delays("month")
})  # stop_db_capturing()
```
This would limit the number of places where code needs to change between
recording and replay, but would still be harder to debug.

##### Data format

I find that using an R structure to store the connection info and the responses,
although natural to the R environment, makes it slightly harder to see changes
in the data when tracking it with git. I wonder if responses shouldn't be `yaml`
or `csv` files, just to make them easier to track with git (to explicitly see a
line change).

Also, since the fixtures are `.R` files, they go through the linter, which
raises errors. It is probably possible to configure, though.

#### Specific comments

Overall, I find the code clear and well commented. The package adheres to the
widespread tidyverse style guide. It makes the code easy to read and welcoming
to new contributors.

--
I wondered why is `.db_mock_path()` prefixed with a `.`?

--
From the doc:
> `start_capturing()` and `stop_capturing()` do the exact same thing as
`start_db_capturing()` and `stop_db_capturing()`. They are both deprecated so
as not to clash with other packages when loaded.

I suggest they should use `.Deprecated`. and since the package is still young,
maybe they could be completely removed?

--
In the `README`:
> dittodb (formerly dbtest)

What's the relation between `rstudio/dbtest` and `jonkeane/dittodb`? Maybe it
doesn't even have to be mentionned?

--
I noticed that the `db-setup/postgres-brew.sh` starts with
rm -rf /usr/local/var/postgres
rm -rf /usr/local/bin/psql
```
Couldn't this lead to users deleting important local data?

--
The documentation seems to refer to fixture and snapshots interchangeably, which
I sometimes found confusing.

####

Good job for this package! It is really well crafted and was a pleasure to
review. I hope my comments, especially on the UX, don't come off as too harsh,
because I really like the solution proposed by the package.
