### 2020-04-17 #366 @helenmiller16
Hi, here is my review! I guess @etiennebr beat me by just 2 hours!

  A couple of notes on vignettes.
    * Just wanted to point out if you didn't know: you can rename your "getting started" vignette to `dittodb` and it will show up in a "getting started" tab on your pkgdown site (https://pkgdown.r-lib.org/reference/build_articles.html#special-links).
    * Good idea to include a vignette detailing the nycflights13 data. I think you should include a bit more info there: instructions for how to get it set up locally and a link to the `nycflights13` package.

  * Documentation for the main functions is good (`with_mock_db` and `start_db_capturing`,  `stop_db_capturing`) which are likely the main functions users would see. Some of the other documented  functions need some clean-up. Make sure you have examples for all exported functions that can be run out of the box.
    * I noticed you mentioned `httptest` once or twice in the docs -- not necessary, but seems like it could be appropriate to link to help pages from `httptest` (If you're not aware of this option, look under "Links" in http://r-pkgs.had.co.nz/man.html#dry2).

  * The examples do not run out-of-the-box. The user would first have to set up a test database. I recommend starting examples that rely on a database with `con <- nycflights13_sqlite()`, which works great (that's how I initially ran the examples).
  *  Some exported functions were missing examples.

  * There is a code of conduct, but no contribution guidelines. There are some extra steps required to get databases set up for testing locally. You touched on it in the README, but I think you need more detailed instructions for how to get this set up, either as part of a CONTRIBUTING file, or as a README in the `db-setup` directory.

     * See my notes under `community guidelines` re setting up test databases locally.

---

I had fun reviewing this package, I think it is well-designed and fills an important need in the R community. I haven't worked with manipulating function signatures before or used trace to add funtion side-effects. Super cool -- works seamlessly, and I learned something! The main places where it still needs some work are around documentation and some code clean-up. I also think it would make sense to tweak the API a bit. I put details in the bullets above or in comments below.

#### API

I'm not really familiar with the `httptest` package, but it looks like you're pretty closely emulating that API. The way that users make database queries vs http requests is different, and I think `dittodb` could better reflect that. When working with databases, users create a connection object which they use to send queries to the database. I think you could more effectively mock that API by creating a mock connection with `dbMockConnect` and then passing that into functions, instead of wrapping everything in `with_mock_db`. This would also make it much easier to run and debug code line-by-line. You're already all set up to do this (I was able to get it to work via `con <- with_mock_db(connectDB(...))`), so it wouldn't be much work to add. That isn't to say you should get rid of `with_mock_db` -- I can see use cases where that would work better, so you could give users both options.

#### Code style

It sometimes took some digging to find the source for functions I was looking for. There are a couple of things I'd suggest that would make it easier for contributers to navigate the code.

* First, I think the names of some of the files in the `R` directory could be more descriptive and consistent. For example, I think `recording.R` might make more sense as something like `capture-results.R` as it contains source for the `_capturing` functions.

* It would also be helpful to reorganize some of the code within the files. Separate out helpers, internal,and exported functions into different sections, and label sections clearly with comments.

#### Other odds and ends

* You have a bunch of functionality written in around debugging, but this is not documented anywhere. At the very least, document it somewhere so users know they have the option of setting the debug level. Even better, include a function like `setDittodbDebugLevel` so they can do it more easily.

* Use `.Deprecated()` for the functions which you have renamed (`start_capturing`, `stop_capturing`). That is the standard and it communicates to users that they are have moved.

* Using `grep` in `redact_columns()` will potentially cause bugs (eg if someone has a df with two columns: `column` and `other_column`, then `redact_columns("column")` will result in an error because both will be matched.

* As I mentioned above, you need some more documentation around how to get the test databases set up. I ran into a couple of things, which you should be sure to address:
      * The README doesn't specify that the scripts are in the `db-setup` directory, so that took a minute to figure out
      * I was already running a postgres server from my machine on port 5432, so I had to stop it to get the dockerized version to work for testing. I recommend using a port other than the default for the test db.

---

Thanks for your work on this package -- I think it's well-implemented and potentially a very useful tool for people who develop around databases. Let me know if you have any questions or if I missed anything -- this is my first review!

Helen
