### 2020-04-17 #366 @jonkeane
Thank you both for these really comprehensive reviews. I've read through them and both have a number of suggestions and places for improvement that will improve `dittodb` a lot and I learned a few things as well (I didn't know that `.Deprecated()` existed  🤯).
I will read through these in more detail and create issues/start working on those issues for all of the suggestions and changes or respond in more detail over the next week or so.

Thank you so much again for the time you both dedicated to this.
