+++ vignettes/getting-started/nycflights/SELECT-e53189.R
@@ -1,6 +1,0 @@
-structure(list(month = 1:12, mean_delay = c(0x1.13d1e5a46fdcp+5, 
-0x1.0d837f713f91bp+5, 0x1.4492c49ce57bcp+5, 0x1.55eaa80cc1cf3p+5, 
-0x1.4f163c59bf3a3p+5, 0x1.ade7fa6ccb7dp+5, 0x1.af9cb5a5cabccp+5, 
-0x1.3c1a8138c6202p+5, 0x1.3671c4fbc61bp+5, 0x1.d0961ced931dbp+4, 
-0x1.b7c0e577c0e57p+4, 0x1.3dd1671e576e9p+5)), class = "data.frame", row.names = c(NA, 
--12L))
