+++ vignettes/getting-started/nycflights/conInfo-.R
@@ -1,3 +1,0 @@
-list(host = "127.0.0.1", username = "travis", dbname = "nycflights", 
-    con.type = "127.0.0.1 via TCP/IP", db.version = "10.4.12-MariaDB-1:10.4.12+maria~bionic", 
-    port = NA_integer_, protocol.version = 10L, thread.id = 27L)
