+++ vignettes/getting-started/nycflights/SELECT-16d120.R
@@ -1,12 +1,0 @@
-structure(list(day = 1:31, mean_delay = c(0x1.55fd54de872a1p+5, 
-0x1.59e435941788dp+5, 0x1.0f393c4c771f6p+5, 0x1.baeff05b9e3cap+4, 
-0x1.1c90412a948cbp+5, 0x1.e78c809868c81p+4, 0x1.67eea7da1dc83p+5, 
-0x1.c1cf68f8013cfp+5, 0x1.603062afefb64p+5, 0x1.a4022947a53a4p+5, 
-0x1.4ee3566690b4fp+5, 0x1.544397c49e8f1p+5, 0x1.444be7f3df2b9p+5, 
-0x1.13e1aaa03ab5cp+5, 0x1.c2b1018fed867p+4, 0x1.0024cd47277d8p+5, 
-0x1.45ed38b2dce19p+5, 0x1.615182d6f26c8p+5, 0x1.4b4858bf824f1p+5, 
-0x1.ef6ff9df0ef05p+4, 0x1.e34fb5df73593p+4, 0x1.7919380ff74a9p+5, 
-0x1.63e23337f7ce1p+5, 0x1.56d7c319729fap+5, 0x1.593dde5542ad9p+5, 
-0x1.1016cdd7d9ab7p+5, 0x1.5c258f74b99bdp+5, 0x1.90ab3046fb0abp+5, 
-0x1.0850127a6b05p+5, 0x1.4957eb50295fbp+5, 0x1.1078e459ff541p+5
-)), class = "data.frame", row.names = c(NA, -31L))
