### 2020-04-18 61954a8
Ensure that column regexes handle more than one column
