+++ .github/workflows/check-mac.yaml
@@ -32,6 +32,9 @@ jobs:
       - uses: r-lib/actions/setup-tinytex@master
         if: contains(matrix.config.args, 'no-manual') == false
 
+      - name: install libxml2
+        run: brew install libxml2
+
       - name: Cache R packages
         uses: actions/cache@v1
         id: rlibcache
