+++ .github/workflows/check-mac.yaml
@@ -49,6 +49,10 @@ jobs:
       - name: Install R dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'httptest'), type = 'source')" -e "remotes::install_deps(dependencies = TRUE);"
 
+     - name: Install xml2 from github (for devel only \#296)
+        if: matrix.config.os == 'macOS-latest' && matrix.config.r == 'devel'
+        run: Rscript -e "remotes::install_github('r-lib/xml2')"
+
       - name: Build R DB Dependencies on cache miss
         if: steps.rlibcache.outputs.cache-hit != 'true'
         run: Rscript -e "install.packages(c('RMariaDB', 'odbc', 'RPostgreSQL', 'RPostgres'), type = 'source');"
