### 2020-04-19 777f1c0
Update redact docs
