+++ man/dot-dittodb_env.Rd
@@ -4,7 +4,9 @@
 \name{.dittodb_env}
 \alias{.dittodb_env}
 \title{an environment for dittodb storing state}
-\format{An object of class \code{environment} of length 0.}
+\format{
+An object of class \code{environment} of length 0.
+}
 \usage{
 .dittodb_env
 }
