### 2020-04-23 2382e17
Rename the getting started vignette
