+++ .github/ISSUE_TEMPLATE.md
@@ -1,10 +1,12 @@
-Please briefly describe your problem and what output you expect. If you have a question, please don't use this form. Instead, ask on <https://stackoverflow.com/> or <https://community.rstudio.com/>.
+Please briefly describe your problem and what output you expect. If you have a more general question, you might find a quicker answer on <https://community.rstudio.com/> or <https://stackoverflow.com/>.
 
-Please include a minimal reproducible example (AKA a reprex). If you've never heard of a [reprex](https://reprex.tidyverse.org/) before, start by reading <https://www.tidyverse.org/help/#reprex>.
+Please include a minimal reproducible example (also called a reprex). If you've never heard of a [reprex](https://reprex.tidyverse.org/) before, there's a great overview at <https://www.tidyverse.org/help/#reprex>.
 
 ---
 
-Brief description of the problem
+### Brief description of the problem
+
+### The kind of database backend you are trying to test
 
 ```r
 # insert reprex here
