+++ .github/workflows/check-windows.yaml
@@ -12,7 +12,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: windows-latest, r: 'release', args: "--no-manual" }
+        - { os: windows-latest, r: '3.6', args: "--no-manual" }
 
     env:
       CRAN: ${{ matrix.config.cran }}
