+++ .github/workflows/check-mac.yaml
@@ -12,7 +12,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: macOS-latest, r: 'release' }
+        - { os: macOS-latest, r: '3.6' }
         - { os: macOS-latest, r: 'devel', args: "--no-manual" }
 
     env:
@@ -70,7 +70,7 @@ jobs:
           path: check
 
       - name: Test coverage
-        if: contains(matrix.config.r, 'release')
+        if: contains(matrix.config.r, '3.6')
         run: |
           # wipe databases
           psql -h 127.0.0.1 -U postgres -f db-setup/postgres-reset.sql
@@ -88,5 +88,5 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: linting
-        if: contains(matrix.config.r, 'release')
+        if: contains(matrix.config.r, '3.6')
         run: Rscript -e "lintr::lint_package()"
