### 2020-04-24 4efa070
Redact bug (#73)
Ensure that column regexes handle more than one column, but by default only matches full columns without adding in specific regex wildcards at the beginning or end
