+++ DESCRIPTION
@@ -51,8 +51,7 @@ Suggests:
     withr,
     rmarkdown,
     lintr
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
 Roxygen: list(markdown = TRUE)
 Language: en-US
 VignetteBuilder: knitr
- 
