+++ .github/workflows/check-linux-ubuntu.yaml
@@ -15,7 +15,7 @@ jobs:
         - { os: ubuntu-18.04, r: '3.3', args: "--no-manual" }
         - { os: ubuntu-18.04, r: '3.4', args: "--no-manual" }
         - { os: ubuntu-18.04, r: '3.5', args: "--no-manual" }
-        - { os: ubuntu-18.04, r: 'release', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '3.6', args: "--no-manual" }
 
     env:
       CRAN: ${{ matrix.config.cran }}
