+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.7.3
 pkgdown: 1.4.1.9000
 pkgdown_sha: 732fb5b9ccc0ef87a73dfd0772ae661a4039b2a8
 articles:
-  getting-started: getting-started.html
+  dittodb: dittodb.html
   nycflights: nycflights.html
   travelling: travelling.html
 urls:
