+++ db-setup/mariadb-docker.sh
@@ -4,6 +4,6 @@
 docker pull mariadb
 
 # on macos, you can only run port forwards
-docker run --name dittodb-mariadb -p 3306:3306 -e MYSQL_ROOT_PASSWORD=r2N5y7V* -d mariadb:latest
+docker run --name dittodb-mariadb -p 6033:3306 -e MYSQL_ROOT_PASSWORD=r2N5y7V* -d mariadb:latest
 
 
