+++ README.md
@@ -1,4 +1,5 @@
 # dittodb
+
 <!-- badges: start -->
 [![Windows](https://github.com/jonkeane/dittodb/workflows/check-windows/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-windows)
 [![Mac](https://github.com/jonkeane/dittodb/workflows/check-mac/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-mac)
@@ -8,12 +9,12 @@
 [![CRAN status](https://www.r-pkg.org/badges/version/dittodb)](https://CRAN.R-project.org/package=dittodb)
 <!-- badges: end -->
 
-
 dittodb (formerly dbtest) is a package that makes testing against databases easy. When writing code that relies on interactions with databases, testing has been difficult without recreating test databases in your CI environment, or resorting to using SQLite databases instead of the database engines you have in production. Both have their downsides: recreating database infrastructure is slow, error prone, and hard to iterate with. Using SQLite works well, right up until you use a feature (like [a full outer join](https://www.sqlite.org/omitted.html)) or has [quirks](https://www.sqlite.org/quirks.html) that might differ from your production database. dittodb solves this by recording database interactions, saving them as mocks, and then replaying them seamlessly during testing. This means that if you can get a query from your database, you can record the response and reliably reproduce that response in tests.
 
 dittodb is heavily inspired by [httptest](https://CRAN.R-project.org/package=httptest), if you've used httptest before, you'll find many of the interactions similar.
 
 ## A quick example
+
 Say we have a Postgres database with some [`nycflights`](https://CRAN.R-project.org/package=nycflights13) data in it and we are writing functions that query this data that we want to test.
 
 For example, we have the simple function that retrieves one airline:
@@ -22,7 +23,6 @@ For example, we have the simple function that retrieves one airline:
 get_an_airline <- function(con) {
   return(dbGetQuery(con, "SELECT carrier, name FROM airlines LIMIT 1"))
 }
-
 ```
 
 But we want to make sure that this function returns what we expect. To do this, we first record the response we get from the production database:
@@ -68,7 +68,6 @@ with_mock_db({
 ```
 
 All without having to ever set a database up on Travis 🎉
-
 
 Alternatively, any other driver could be used:
 ```r
@@ -124,6 +123,18 @@ In order to test `dittodb` recording functionality locally or on CI, it is helpf
 To get local databases, the easiest way is to use docker and run either the `postgres-docker-reset.sh` or `mariadb-docker-reset.sh` which will pull a docker image and set up a test database with the user and passwords that the `dittodb` tests are expecting (and will stop and remove the docker images if they are present). 
 
 On continuous integration, (using GitHub Actions) these scripts in the `db-setup` directory are used to set up these test databases so we can run integration tests (predominantly in the file `tests/testthat/test-dbi-generic-integration.R`).
+
+## Local tests
+
+Database configuration scripts are located in `db-setup`.
+
+## PostgreSQL
+
+You can use `db-setup/postgres-docker.sh` and ``tests/test-dbi-generic-integration.R`/test-nycflights13.R` directly, the setup uses port `2345` to avoid conflicts with (anyexisting PostgreSQL instances
+
+## PostgreSQL
+
+You can use `db-setup/mariadb-docker.sh` aand `tests/test-dbi-generic-integration.R` irectly, the setup uses port `6033` to avoid conflicts with (anyexisting PmariaDBinstances
 
 ## Code of Conduct
 
