+++ db-setup/populate-dbs.sh
@@ -1,5 +1,5 @@
 psql -h 127.0.0.1 -U postgres -p 2345 -f db-setup/postgres-reset.sql
 psql -h 127.0.0.1 -U postgres -p 2345 -d nycflights -f db-setup/postgres-nycflights.sql
 
-mysql -h 127.0.0.1 -u root -p 6033 -e "source db-setup/mariadb-reset.sql"
-mysql -h 127.0.0.1 -u root -p 6033 -e "source db-setup/mariadb-nycflights.sql"
+mysql -h 127.0.0.1 -u root --port=6033 -e "source db-setup/mariadb-reset.sql"
+mysql -h 127.0.0.1 -u root --port=6033 -e "source db-setup/mariadb-nycflights.sql"
