+++ db-setup/mariadb-brew.sh
@@ -4,4 +4,5 @@ brew services start mariadb
 mkdir /usr/local/etc/my.cnf.d
 sleep 2
 # TODO: add dbrootpass if it doesn't exist?
-sudo mysql -u root -e "ALTER USER 'root'@'localhost' IDENTIFIED BY '$DBROOTPASS';"
+sudo mysqld --port=6033
+sudo mysql -u root --port=6033 -e "ALTER USER 'root'@'localhost' IDENTIFIED BY '$DBROOTPASS';"
