+++ .github/workflows/check-mac.yaml
@@ -77,12 +77,12 @@ jobs:
         if: contains(matrix.config.r, '3.6')
         run: |
           # wipe databases
-          psql -h 127.0.0.1 -U postgres -f db-setup/postgres-reset.sql
-          mysql -h 127.0.0.1 -u root -e "source db-setup/mariadb-reset.sql"
+          psql -h 127.0.0.1 -U postgres -p 2345 -f db-setup/postgres-reset.sql
+          mysql -h 127.0.0.1 -u root --port=6033 -e "source db-setup/mariadb-reset.sql"
 
           # restore
-          psql -h 127.0.0.1 -U postgres -d nycflights -f db-setup/postgres-nycflights.sql
-          mysql -h 127.0.0.1 -u root -e "source db-setup/mariadb-nycflights.sql"
+          psql -h 127.0.0.1 -U postgres -p 2345 -d nycflights -f db-setup/postgres-nycflights.sql
+          mysql -h 127.0.0.1 -u root --port=6033 -e "source db-setup/mariadb-nycflights.sql"
 
           # coverage
           Rscript -e 'remotes::install_github("r-lib/covr@gh-actions")'
