+++ db-setup/postgres-brew.sh
@@ -14,6 +14,7 @@ EOT
 initdb /usr/local/var/postgres
 pg_ctl -D /usr/local/var/postgres start
 /usr/local/opt/postgres/bin/createuser -s postgres
-psql -U postgres -c 'SHOW config_file'
-# sed -i 's/5432/2345/g' foo
+# psql -U postgres -c 'SHOW config_file'
+sed -i 's/5432/2345/g' /usr/local/var/postgres/postgresql.conf
+pg_ctl -D /usr/local/var/postgres restart
 sleep 2
