+++ db-setup/postgres-brew.sh
@@ -15,6 +15,6 @@ initdb /usr/local/var/postgres
 pg_ctl -D /usr/local/var/postgres start
 /usr/local/opt/postgres/bin/createuser -s postgres
 # psql -U postgres -c 'SHOW config_file'
-sed -i 's/5432/2345/g' /usr/local/var/postgres/postgresql.conf
+sed -i 's/#port = 5432/port = 2345/g' /usr/local/var/postgres/postgresql.conf
 pg_ctl -D /usr/local/var/postgres restart
 sleep 2
