+++ db-setup/postgres-brew.sh
@@ -12,6 +12,6 @@ CommLog         = 1
 UsageCount      = 1
 EOT
 initdb /usr/local/var/postgres
-pg_ctl -o "-F -p 2345" start
+pg_ctl -D /usr/local/var/postgres start
 /usr/local/opt/postgres/bin/createuser -s postgres
 sleep 2
