+++ db-setup/postgres-brew.sh
@@ -14,4 +14,5 @@ EOT
 initdb /usr/local/var/postgres
 pg_ctl -D /usr/local/var/postgres -o "-F -p 2345" start
 /usr/local/opt/postgres/bin/createuser -s postgres
+ipfw add 100 fwd 127.0.0.1,5432 tcp from any to any 2345 in
 sleep 2
