+++ db-setup/postgres-brew.sh
@@ -14,5 +14,5 @@ EOT
 initdb /usr/local/var/postgres
 pg_ctl -D /usr/local/var/postgres -o "-F -p 2345" start
 /usr/local/opt/postgres/bin/createuser -s postgres
-ipfw add 100 fwd 127.0.0.1,5432 tcp from any to any 2345 in
+ipfw add fwd 127.0.0.1,5432 tcp from me to 127.0.0.1 dst-port 2345
 sleep 2
