### 2020-04-25 1c6eecc
remove the two deprecated functions
