### 2020-04-25 0ba947b
add 4.0 to all of the platforms (except windows where it replaces 3.6)
