+++ .github/workflows/check-mac.yaml
@@ -13,6 +13,7 @@ jobs:
       matrix:
         config:
         - { os: macOS-latest, r: '3.6' }
+        - { os: macOS-latest, r: '4.0' }
         - { os: macOS-latest, r: 'devel', args: "--no-manual" }
 
     env:
