+++ .github/workflows/check-mac.yaml
@@ -75,7 +75,7 @@ jobs:
           path: check
 
       - name: Test coverage
-        if: contains(matrix.config.r, '3.6')
+        if: contains(matrix.config.r, '4.0')
         run: |
           # wipe databases
           psql -h 127.0.0.1 -U postgres -f db-setup/postgres-reset.sql
@@ -93,5 +93,5 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: linting
-        if: contains(matrix.config.r, '3.6')
+        if: contains(matrix.config.r, '4.0')
         run: Rscript -e "lintr::lint_package()"
