### 2020-04-25 9228590
R 4.0 in CI (#78)
* add 4.0 to all of the platforms where it is available 
* No need for special xml2 anymore
