+++ .github/workflows/check-mac.yaml
@@ -12,7 +12,8 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: macOS-latest, r: '3.6' }
+        - { os: macOS-latest, r: '3.6', args: "--no-manual" }
+        - { os: macOS-latest, r: '4.0' }
         - { os: macOS-latest, r: 'devel', args: "--no-manual" }
 
     env:
@@ -46,10 +47,6 @@ jobs:
            bash db-setup/mariadb-brew.sh
         if: contains(runner.os, 'macOS')
 
-      - name: Install xml2 from github (for devel only \#296)
-        if: matrix.config.os == 'macOS-latest' && matrix.config.r == 'devel'
-        run: Rscript -e "install.packages('remotes', type = 'source')" -e "remotes::install_github('r-lib/xml2')"
-
       - name: Install R dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'httptest'), type = 'source')" -e "remotes::install_deps(dependencies = TRUE);"
 
@@ -74,7 +71,7 @@ jobs:
           path: check
 
       - name: Test coverage
-        if: contains(matrix.config.r, '3.6')
+        if: contains(matrix.config.r, '4.0')
         run: |
           # wipe databases
           psql -h 127.0.0.1 -U postgres -f db-setup/postgres-reset.sql
@@ -92,5 +89,5 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: linting
-        if: contains(matrix.config.r, '3.6')
+        if: contains(matrix.config.r, '4.0')
         run: Rscript -e "lintr::lint_package()"
