+++ R/mock-paths.R
@@ -2,12 +2,12 @@
 #'
 #' By default, `with_mock_api` will look for mocks relative to the current
 #' working directory (or the test directory). If you want to look in other
-#' places, you can call `.db_mock_paths` to add directories to the search path.
+#' places, you can call `db_mock_paths` to add directories to the search path.
 #'
 #' It works like [base::.libPaths()]: any directories you specify will be added
 #' to the list and searched first. The default directory will be searched last.
 #' Only unique values are kept: if you provide a path that is already found in
-#' `.db_mock_paths`, the result effectively moves that path to the first
+#' `db_mock_paths`, the result effectively moves that path to the first
 #' position.
 #'
 #' For finer-grained control, or to completely override the default behavior
@@ -17,20 +17,23 @@
 #' This function is similar to `.mockPaths()` from
 #' [httptest](https://CRAN.R-project.org/package=httptest)
 #'
+#' The function `.db_mock_paths` is the same as `db_mock_paths` although it is
+#' deprecated and should not be used.
+#'
 #' @param new Either a character vector of path(s) to add, or `NULL` to reset
 #' to the default.
 #' @return If `new` is omitted, the function returns the current search paths, a
 #' character vector. If `new` is provided, the updated value will be returned
 #' invisibly.
 #' @examples
-#' identical(.db_mock_paths(), c("tests/testthat/", "."))
-#' .db_mock_paths("/var/somewhere/else")
-#' identical(.db_mock_paths(), c("/var/somewhere/else", "tests/testthat/", "."))
-#' .db_mock_paths(NULL)
-#' identical(.db_mock_paths(), c("tests/testthat/", "."))
-#' @rdname mockPaths
+#' identical(db_mock_paths(), c("tests/testthat/", "."))
+#' db_mock_paths("/var/somewhere/else")
+#' identical(db_mock_paths(), c("/var/somewhere/else", "tests/testthat/", "."))
+#' db_mock_paths(NULL)
+#' identical(db_mock_paths(), c("tests/testthat/", "."))
+#' @name mockPaths
 #' @export
-.db_mock_paths <- function(new) {
+db_mock_paths <- function(new) {
   # use both "." and testthat::test_path(".") in case they are different
   def <- unique(c("tests/testthat/", "."))
 
@@ -52,8 +55,12 @@
 
 # for backwards compatibility
 #' @export
+#' @rdname mockPaths
 #' @keywords internal
-.mockPaths <- .db_mock_paths
+.db_mock_paths <- function(new) {
+  .Deprecated("db_mock_paths")
+  return(db_mock_paths(new))
+}
 
 #' Run the DBI queries in an alternate mock directory
 #'
@@ -69,12 +76,12 @@
 #' @return nothing
 #' @export
 with_mock_path <- function(path, expr, replace = FALSE) {
-  oldmp <- .db_mock_paths()
+  oldmp <- db_mock_paths()
   if (replace) {
     options(dittodb.mock.paths = path)
   } else {
     ## Append
-    .db_mock_paths(path)
+    db_mock_paths(path)
   }
   on.exit(options(dittodb.mock.paths = oldmp))
   return(eval.parent(expr))
