+++ NAMESPACE
@@ -2,7 +2,7 @@
 
 export(.db_mock_paths)
 export(.dittodb_env)
-export(.mockPaths)
+export(db_mock_paths)
 export(db_path_sanitize)
 export(dittodb_debug_level)
 export(get_dbname)
