### 2020-04-26 a9c7475
Add functionality to start and stop dittodb mock usage
