+++ tests/testthat/setup.R
@@ -11,13 +11,8 @@ db_user <- "travis"
 db_pass <- "*V7y5N2r"
 
 
-windows <- any(
-  .Platform$OS.type == "windows",
-  tolower(Sys.info()['sysname']) == "windows",
-  na.rm = TRUE
-)
-
-odbc_driver <- if (windows) {
+win_x64 <- .Platform$OS.type == "windows" & Sys.info()[["machine"]] == "x86_64"
+odbc_driver <- if (win_x64) {
   "PostgreSQL Unicode(x64)"
 } else {
   "PostgreSQL Unicode"
