### 2020-04-26 1c70281
a different way of detecting architecture?
