+++ tests/testthat/setup.R
@@ -11,7 +11,7 @@ db_user <- "travis"
 db_pass <- "*V7y5N2r"
 
 
-win_x64 <- .Platform$OS.type == "windows" & Sys.info()[["machine"]] == "x86_64"
+win_x64 <- .Platform$OS.type == "windows" & version$arch == "x86_64"
 odbc_driver <- if (win_x64) {
   "PostgreSQL Unicode(x64)"
 } else {
