### 2020-04-26 1aa971e
install the x86 version too?
