### 2020-04-26 571bc0a
skip odbc 32-bit windows
