### 2020-04-26 9ef4118
enable tests on CI
