+++ .github/workflows/check-mac.yaml
@@ -20,6 +20,8 @@ jobs:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       CRAN: ${{ matrix.config.cran }}
       DBROOTPASS: r2N5y7V*
+      DITTODB_ENABLE_PG: true
+      DITTODB_ENABLE_MARIA: true
 
     steps:
       - uses: actions/checkout@v1
