+++ tests/testthat/test-nycflights13.R
@@ -1,7 +1,6 @@
-# sqlite ----
-
 context("nycflights13 writing functions")
 
+# sqlite ----
 test_that("nycflights sqlite can be created", {
   # with DBI
   con <- nycflights13_sqlite()
@@ -13,13 +12,8 @@ test_that("nycflights sqlite can be created", {
   dbDisconnect(con)
 })
 
-
-skip_locally("use postgres-docker.sh and then this can be unskipped locally")
-
 # psql ----
-
 # odbc ----
-
 test_that("odbc", {
   skip_env("odbc")
 
@@ -60,40 +54,43 @@ test_that("odbc", {
 })
 
 # rpostgresql ----
+test_that("RPostgreSQL", {
+  skip_env("RPostgreSQL")
 
-con_rpostgresql <- RPostgreSQL::dbConnect(
-  drv = DBI::dbDriver("PostgreSQL"),
-  host = "127.0.0.1",
-  dbname = "postgres",
-  user = db_user,
-  password = db_pass,
-  port = 5432
-)
+  con_rpostgresql <- RPostgreSQL::dbConnect(
+    drv = DBI::dbDriver("PostgreSQL"),
+    host = "127.0.0.1",
+    dbname = "postgres",
+    user = db_user,
+    password = db_pass,
+    port = 5432
+  )
 
-test_that("DBI, with a new schema creation and rpostgresql package", {
-  expect_message(
-    nycflights13_sql(con_rpostgresql, schema = "new_schema_rpostgresql"),
-    paste0(c(
-      "Creating the testing database from nycflights13",
-      "Creating table: airlines",
-      "Creating table: airports",
-      "Creating table: flights",
-      "Creating table: planes",
-      "Creating table: weather"
-    ),
-    collapse = "|"
+  test_that("DBI, with a new schema creation and rpostgresql package", {
+    expect_message(
+      nycflights13_sql(con_rpostgresql, schema = "new_schema_rpostgresql"),
+      paste0(c(
+        "Creating the testing database from nycflights13",
+        "Creating table: airlines",
+        "Creating table: airports",
+        "Creating table: flights",
+        "Creating table: planes",
+        "Creating table: weather"
+      ),
+      collapse = "|"
+      )
     )
-  )
-})
+  })
 
-test_that("DBI, with a same schema creation and rpostgresql package", {
-  expect_message(
-    nycflights13_sql(con_rpostgresql, schema = "new_schema_rpostgresql"),
-    "Creating the testing database from nycflights13"
-  )
-})
+  test_that("DBI, with a same schema creation and rpostgresql package", {
+    expect_message(
+      nycflights13_sql(con_rpostgresql, schema = "new_schema_rpostgresql"),
+      "Creating the testing database from nycflights13"
+    )
+  })
 
-dbDisconnect(con_rpostgresql)
+  dbDisconnect(con_rpostgresql)
+})
 
 # rpostgres ----
 test_that("RPostgres", {
