### 2020-04-26 a7f4692
More comments, a more specific architecture for windows
