+++ tests/testthat/helpers.R
@@ -1,17 +1,19 @@
 skip_env <- function(which = c("RPostgres", "RMariaDB", "RPostgreSQL", "odbc")) {
   if (any(c("RPostgres", "RPostgreSQL", "odbc") %in% which) &&
-      tolower(Sys.getenv("DITTODB_ENABLE_PG")) != "true") {
+      tolower(Sys.getenv("DITTODB_ENABLE_PG_TESTS")) != "true") {
     skip("Skipping tests that need functioning Postgres.")
   }
 
-  # installing 32-bit odbc drivers on windows is hard, so skip on CI
-  win_x32 <- .Platform$OS.type == "windows" && version$arch != "x86_64"
+  # installing both 32-bit and 64-bit odbc drivers on windows is non-trivial, so
+  # skipping the 32-bit version of odbc tests which struggle to find an
+  # appropriate 32-bit driver.
+  win_x32 <- .Platform$OS.type == "windows" && version$arch == "x86_32"
   if ("odbc" %in% which && win_x32) {
     skip("Skipping odbc tests on 32bit windows.")
   }
 
   if ("RMariaDB" %in% which &&
-      tolower(Sys.getenv("DITTODB_ENABLE_MARIA")) != "true") {
+      tolower(Sys.getenv("DITTODB_ENABLE_MARIA_TESTS")) != "true") {
     skip("Skipping tests that need functioning MariaDB.")
   }
 
