### 2020-04-26 37870d2
Some more cleanup
