### 2020-04-26 023689a
Add functionality to start and stop dittodb mock usage (#80)
Add functionality to start and stop dittodb mock usage
