+++ .github/workflows/check-mac.yaml
@@ -44,7 +44,8 @@ jobs:
 
       - name: Install databases (macOS)
         run: |
-           brew update
+           brew update || brew update
+           brew upgrade
            bash db-setup/postgres-brew.sh
            bash db-setup/mariadb-brew.sh
         if: contains(runner.os, 'macOS')
