### 2020-04-26 c6d9973
Get C code registration right for once
