+++ vignettes/developing-dittodb.Rmd
@@ -21,12 +21,11 @@ knitr::opts_chunk$set(
 db_mock_paths("dittodb")
 ```
 
-We welcome conributions from anyone, no matter how small or trivial. Documentation additions or typo fixes are especially welcome. For larger, more-functional changes, either see if there is an issue open on Github already, or open one with a proposal of the change you would like to make. Please also see our [code of conduct](../docs/CODE_OF_CONDUCT.html) and [contributing guide](../docs/CONTRIBUTING.html).
+We welcome contributions from anyone, no matter how small or trivial. Documentation additions or typo fixes are especially welcome. For larger, more-functional changes, either see if there is an issue open on Github already, or open one with a proposal of the change you would like to make. Please also see our [code of conduct](../docs/CODE_OF_CONDUCT.html) and [contributing guide](../docs/CONTRIBUTING.html).
 
 Developing `dittodb` requires is a bit more complication than developing other R packages for a few reasons: 1. setting up all of the databases to fully test recording is complicated (which is in some ways the exact reason `dittodb` exists, so you don't have to do this!) and 2. some of the mechanisms that make `dittodb` work aren't commonly used in other R packages.
 
 ## Setting up databases
-
 In order to fully test that `dittodb` works, we aim to have full coverage and test as many database backends as possible for both recording and using as a mocked database. To do this on CI (or even locally) can be finicky to get working (and on the CI front, we did it once, so that you can use `dittodb` and you won't have to setup your own database backend just to run tests!). We include in the repository a few scripts that make it (relatively) easy to setup testing database backends, as well as some scripts that we use to setup database backends on Github Actions.
 
 ### What we test
@@ -39,31 +38,30 @@ We currently test against the following database backends with Github Actions fo
 ### How to setup test databases locally
 All of these (with the exception of SQLite) are tested in the test file `test-dbi-generic-integration.R`. However, tests for each database are only run if specific environment variables are set that trigger them. The reason for this is so that it is easy to test locally without needing to setup databases, but we are covered by these tests being run on Github Actions. If you would like to run these tests locally, you can set the following environment variables and run tests as usual (e.g. with `R CMD check`, `devtools::check()`, `devtools::test()`)
 
-* if `DITTODB_ENABLE_PG_TESTS` is `TRUE`, then postgres-based tests will be run
-* if `DITTODB_ENABLE_MARIA_TESTS` is `TRUE`, then mariadb-based tests will be run
+* if `DITTODB_ENABLE_PG_TESTS` is `TRUE`, then Postgres-based tests will be run
+* if `DITTODB_ENABLE_MARIA_TESTS` is `TRUE`, then MariaDB-based tests will be run
 
-There are a few scripts included in the `db-setup` folder that are helpful for setting up databses. For local tests, we highly recommend using the docker scripts:
+There are a few scripts included in the `db-setup` folder that are helpful for setting up databases. For local tests, we highly recommend using the docker scripts:
 
-* `db-setup/mariadb-docker.sh` and `db-setup/mariadb-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs maridb in that container (running on the default port 3306), and loads the correct test user and test data into the database for running tests.
-* `db-setup/postgres-docker.sh` and `db-setup/postgres-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs postgresql in that container (running on the default port 5432), and loads the correct test user and test data into the database for running tests.
+* `db-setup/mariadb-docker.sh` and `db-setup/mariadb-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs MariaDB in that container (running on the default port 3306), and loads the correct test user and test data into the database for running tests.
[…]
