### 2020-04-26 94840d2
A few minor cleanups (#81)
