+++ NAMESPACE
@@ -53,5 +53,3 @@ importFrom(testthat,with_mock)
 importFrom(utils,head)
 importFrom(utils,installed.packages)
 useDynLib(dittodb, .registration = TRUE)
-useDynLib(dittodb,duplicate_)
-useDynLib(dittodb,reassign_function)
