+++ db-setup/postgres-docker.sh
@@ -3,5 +3,8 @@
 
 docker pull postgres
 
+# Get the port used in DITTODB_MARIA_TEST_PORT, else use the default
+POSTGRESPORT=${DITTODB_PG_TEST_PORT:-5432}
+
 # on macos, you can only run port forwards
-docker run --name dittodb-postgres -p 5432:5432 -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=r2N5y7V* -d postgres:latest
+docker run --name dittodb-postgres -p $POSTGRESPORT:5432 -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=r2N5y7V* -d postgres:latest
