+++ vignettes/developing-dittodb.Rmd
@@ -46,7 +46,7 @@ There are a few scripts included in the `db-setup` folder that are helpful for s
 * `db-setup/mariadb-docker.sh` and `db-setup/mariadb-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs MariaDB in that container (running on the default port 3306), and loads the correct test user and test data into the database for running tests.
 * `db-setup/postgres-docker.sh` and `db-setup/postgres-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs Postgres in that container (running on the default port 5432), and loads the correct test user and test data into the database for running tests.
 
-If you've already got databases running on the default ports (3306 for MariaDB and 5432 for Postgres) and you want to use the docker scripts, we recommend that you change the ports that docker is using for each and then you use the `DITTODB_MARIA_TEST_PORT` and `DITTODB_PG_TEST_PORT` environment variables to change which port `dittodb` uses to connect to the test databases.
+If you've already got databases running on the default ports (3306 for MariaDB and 5432 for Postgres) and you want to use the docker scripts, we recommend that you change the ports that docker is using for each and then you use the `DITTODB_MARIA_TEST_PORT` and `DITTODB_PG_TEST_PORT` environment variables to change which port `dittodb` uses to connect to the test databases. The docker scripts above will use these environment variables to map ports if they are set (and exported) for convenience. One thing to note: during `dittodb` tests, if some database drivers attempt to connect to not-running or on-the-wrong-port database backends, they can segfault instead of erroring with a more informative error. If you see this, the first thing to check is that the port variables are being set correctly and that the database backend is up and running normally.
 
 Both of these utilize a few SQL scripts for their respective backends. These might be useful if you're manually adding the test data into a database you already have running, but if you're using the docker scripts above, you shouldn't need to use them at all. 
 
@@ -54,7 +54,7 @@ Both of these utilize a few SQL scripts for their respective backends. These mig
 * `db-setup/[mariadb|postgres]-nycflights.sql` creates the necessary tables in the `nycflights` database for use in testing.
 * `db-setup/populate-dbs.sh` uses the above scripts to populate the databases on Github Actions.
 
-### ☠️ What not to touch ☠️
+### ☠️ What not to run ☠️
 The other scripts (e.g. `db-setup/[mariadb|postgres]-brew.sh`) are only intended for use on Github Actions and should not be run locally. They include commands that will remove files necessary to reset database setups that allow for tests to be run. Running them locally will delete files that you might care about.
 
 ## Some of the tricky bits that `dittodb` uses
