### 2020-04-28 3c0fffe
Move some of the scripts around.
