+++ vignettes/developing-dittodb.Rmd
@@ -43,8 +43,8 @@ All of these (with the exception of SQLite) are tested in the test file `test-db
 
 There are a few scripts included in the `db-setup` folder that are helpful for setting up databases. For local tests, we highly recommend using the docker scripts:
 
-* `db-setup/mariadb-docker.sh` and `db-setup/mariadb-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs MariaDB in that container (running on the default port 3306), and loads the correct test user and test data into the database for running tests.
-* `db-setup/postgres-docker.sh` and `db-setup/postgres-docker-reset.sh` which starts (or stops and then starts for the `*-reset.sh` script) a docker container, installs Postgres in that container (running on the default port 5432), and loads the correct test user and test data into the database for running tests.
+* `db-setup/local-mariadb-docker-setup.sh` which starts (or stops and then starts if it's already running) a docker container, installs MariaDB in that container (running on the default port 3306), and loads the correct test user and test data into the database for running tests.
+* `db-setup/local-postgres-docker-setup.sh` which starts (or stops and then starts if it's already running) a docker container, installs Postgres in that container (running on the default port 5432), and loads the correct test user and test data into the database for running tests.
 
 If you've already got databases running on the default ports (3306 for MariaDB and 5432 for Postgres) and you want to use the docker scripts, we recommend that you change the ports that docker is using for each and then you use the `DITTODB_MARIA_TEST_PORT` and `DITTODB_PG_TEST_PORT` environment variables to change which port `dittodb` uses to connect to the test databases. The docker scripts above will use these environment variables to map ports if they are set (and exported) for convenience. One thing to note: during `dittodb` tests, if some database drivers attempt to connect to not-running or on-the-wrong-port database backends, they can segfault instead of erroring with a more informative error. If you see this, the first thing to check is that the port variables are being set correctly and that the database backend is up and running normally.
 
@@ -55,7 +55,7 @@ Both of these utilize a few SQL scripts for their respective backends. These mig
 * `db-setup/populate-dbs.sh` uses the above scripts to populate the databases on Github Actions.
 
 ### ☠️ What not to run ☠️
-The other scripts (e.g. `db-setup/[mariadb|postgres]-brew.sh`) are only intended for use on Github Actions and should not be run locally. They include commands that will remove files necessary to reset database setups that allow for tests to be run. Running them locally will delete files that you might care about.
+The other scripts (e.g. `db-setup/[mariadb|postgres]-brew.sh` and `db-setup/[mariadb|postgres]-docker-container-only.sh`) are only intended for use on Github Actions and should not be run locally. They include commands that will remove files necessary to reset database setups that allow for tests to be run. Running them locally will delete files that you might care about.
 
 ## Some of the tricky bits that `dittodb` uses
 In order to provide a seamless experience between using a real database connection and using the mocked version of the database `dittodb` uses some features of R that are pretty uncommon. This is not intended to be a comprehensive description of `dittodb`'s architecture, but a few things that are uncommon or a little strange.
