+++ tests/testthat/test-nycflights13.R
@@ -24,7 +24,7 @@ test_that("odbc", {
     Database = "postgres",
     UID = db_user,
     PWD = db_pass,
-    Port = 5432
+    Port = testing_port("postgres")
   )
 
   test_that("DBI, with a new schema creation and odbc package", {
@@ -63,7 +63,7 @@ test_that("RPostgreSQL", {
     dbname = "postgres",
     user = db_user,
     password = db_pass,
-    port = 5432
+    port = testing_port("postgres")
   )
 
   test_that("DBI, with a new schema creation and rpostgresql package", {
@@ -102,7 +102,7 @@ test_that("RPostgres", {
     dbname = "postgres",
     user = db_user,
     password = db_pass,
-    port = 5432
+    port = testing_port("postgres")
   )
 
   test_that("DBI, with a new schema creation and rpostgres package", {
