+++ db-setup/mariadb-docker.sh
@@ -3,7 +3,10 @@
 
 docker pull mariadb
 
+# Get the port used in DITTODB_MARIA_TEST_PORT, else use the default
+MARIAPORT=${DITTODB_MARIA_TEST_PORT:-3306}
+
 # on macos, you can only run port forwards
-docker run --name dittodb-mariadb -p 3306:3306 -e MYSQL_ROOT_PASSWORD=r2N5y7V* -d mariadb:latest
+docker run --name dittodb-mariadb -p $MARIAPORT:3306 -e MYSQL_ROOT_PASSWORD=r2N5y7V* -d mariadb:latest
 
 
