### 2020-05-03 63d0dac
Developer docs (#82)
New developer docs vignette that explains better how to develop `dittodb`
