+++ .github/workflows/check-linux-ubuntu.yaml
@@ -50,9 +50,9 @@ jobs:
 
       - name: Install databases and odbc manual adjustments (linux)
         run: |
-          bash db-setup/postgres-docker.sh
+          bash db-setup/postgres-docker-container-only.sh
           sudo service mysql stop
-          bash db-setup/mariadb-docker.sh
+          bash db-setup/mariadb-docker-container-only.sh
           # the driver as installed in odbcinst.ini isn't located
           sudo sed -i 's/Driver=psqlodbcw.so/Driver=\/usr\/lib\/x86_64-linux-gnu\/odbc\/psqlodbcw.so/g' /etc/odbcinst.ini
           sudo sed -i 's/Driver=psqlodbca.so/Driver=\/usr\/lib\/x86_64-linux-gnu\/odbc\/psqlodbca.so/g' /etc/odbcinst.ini
