### 2020-05-05 3c62bb1
More examples, now more runnable
