### 2020-05-06 464d5d7
More examples, now more runnable (#86)
