+++ R/mock-paths.R
@@ -10,6 +10,11 @@
 #' `db_mock_paths`, the result effectively moves that path to the first
 #' position.
 #'
+#' When you are capturing fixtures (e.g. with [`start_db_capturing`]), the first
+#' path is used as the path to save the fixtures in. For this reason, you may
+#' want to set the `last` argument to `TRUE` if you want to read from a
+#' directory but don't want to write to it.
+#'
 #' For finer-grained control, or to completely override the default behavior
 #' of searching in the current working directory, you can set the option
 #' "dittodb.mock.paths" directly.
@@ -20,20 +25,30 @@
 #' The function `.db_mock_paths` is the same as `db_mock_paths` although it is
 #' deprecated and should not be used.
 #'
-#' @param new Either a character vector of path(s) to add, or `NULL` to reset
+#' @param new either a character vector of path(s) to add, or `NULL` to reset
 #' to the default.
+#' @param last a logical, should the new path given be added to the end of the
+#' list of paths? (default: `FALSE`)
+#'
 #' @return If `new` is omitted, the function returns the current search paths, a
 #' character vector. If `new` is provided, the updated value will be returned
 #' invisibly.
 #' @examples
+#' \dontrun{
+#' # reset mock paths to default
+#' db_mock_paths(NULL)
+#'
 #' identical(db_mock_paths(), c("tests/testthat/", "."))
 #' db_mock_paths("/var/somewhere/else")
 #' identical(db_mock_paths(), c("/var/somewhere/else", "tests/testthat/", "."))
 #' db_mock_paths(NULL)
 #' identical(db_mock_paths(), c("tests/testthat/", "."))
+#' db_mock_paths("/var/somewhere/else", last = TRUE)
+#' identical(db_mock_paths(), c("tests/testthat/", ".", "/var/somewhere/else"))
+#' }
 #' @name mockPaths
 #' @export
-db_mock_paths <- function(new) {
+db_mock_paths <- function(new, last = FALSE) {
   # use both "." and testthat::test_path(".") in case they are different
   def <- unique(c("tests/testthat/", "."))
 
@@ -44,13 +59,19 @@ db_mock_paths <- function(new) {
   } else if (is.null(new)) {
     ## We're calling the function to reset to the default
     options(dittodb.mock.paths = new)
-    return(invisible(current))
   } else {
     ## We're adding one or more paths
-    current <- unique(c(new, current))
+    if (last) {
+      # if we are post-pending
+      current <- unique(c(current, new), fromLast = TRUE)
+    } else {
+      # if we are pre-pending
+      current <- unique(c(new, current))
+    }
     options(dittodb.mock.paths = current)
-    return(invisible(current))
   }
+
+  return(invisible(current))
 }
 
 # for backwards compatibility
