+++ vignettes/dittodb.Rmd
@@ -92,11 +92,11 @@ test_that("mean_delays()", {
 })
 ```
 
-And this works just fine if we only ever run your tests locally, but if we want to [run our tests with a Continuous Integration system](http://r-pkgs.had.co.nz/check.html#check) (and yes, we want to do that!), this won't work without first setting up our production database of flights. For our tests, we don't actually need to connect to the database and get new data (and, in fact, that would make some tests fail erroneously suddenly if the underlying changed). Instead, what we want is to take a snapshot of running the test code, and then be able to use that recorded snapshot when we run tests later.
+And this works just fine if we only ever run your tests locally, but if we want to [run our tests with a Continuous Integration system](http://r-pkgs.had.co.nz/check.html#check) (and yes, we want to do that!), this won't work without first setting up our production database of flights. For our tests, we don't actually need to connect to the database and get new data (and, in fact, that would make some tests fail erroneously suddenly if the underlying changed). Instead, what we want is to take a snapshot of what happens when running the test code, and then be able to use that snapshot when we run tests later. These snapshots are frequently called fixtures (though you might hear people use other names like stubs or mocks).
 
-## Recording snapshots
+## Recording fixtures
 
-We can record snapshots of the database interactions with the commands `start_db_capturing()`, run the functions we want to record, and then stop recording with `stop_db_capturing()`.
+We can record fixtures of the database interactions with the commands `start_db_capturing()`, run the functions we want to record, and then stop recording with `stop_db_capturing()`.
 
 ```{r recording, eval = FALSE}
 start_db_capturing()
@@ -104,11 +104,11 @@ out <- mean_delays("month")
 stop_db_capturing()
 ```
 
-This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-e53189.R` which is the snapshot for this example. This `SELECT-*` file contains the data that was received from the database for use in tests.
+This will write a new folder (by default in `./tests/testthat/`) with the name of the database (here: `nycflights`) and then write one file with the name `SELECT-e53189.R` which is the fixture for this example. This `SELECT-*` file contains the data that was received from the database for use in tests.
 
 ## `with_mock_db()`
 
-Now that we have a snapshot, we can use that snapshot by wrapping our call that includes a database interaction with the function `with_mock_db()`. This will look for snapshots and use those.
+Now that we have a fixture, we can use that fixture by wrapping our call that includes a database interaction with the function `with_mock_db()`. This will look for fixtures and use those.
 
 ```{r with_mock_1}
 with_mock_db(
@@ -131,9 +131,9 @@ with_mock_db(
 )
 ```
 
-When placed inside of `with_mock_db(...)` a call to `mean_delays("month")` will return what we saved as our snapshot _as if it had actually connected to the database_ without needing the database to be installed, reachable, or operational.
+When placed inside of `with_mock_db(...)` a call to `mean_delays("month")` will return what we saved as our fixture _as if it had actually connected to the database_ without needing the database to be installed, reachable, or operational.
 
-If we wanted to test that a day-based aggregation works, we can, although we will have to make a new snapshot. First we would run the following interactively:
+If we wanted to test that a day-based aggregation works, we can, although we will have to make a new fixture. First we would run the following interactively:
 
 ```{r recording_days, eval = FALSE}
 start_db_capturing()
@@ -159,11 +159,11 @@ There are a few things to be careful about when using dittodb.
 
 ### When to call `dbConnect()`
 
[…]
