### 2020-05-06 501fedf
Add docs about debugging, some other clean up
