+++ R/connection.R
@@ -2,7 +2,8 @@
 #'
 #' Various methods (`dbSendQuery`, `dbFetchQuery`) that are mocks of the DBI
 #' methods of the same name. Instead of actually interacting with a database,
-#' they read in mock responses and the code proceeds after that.
+#' they read in mock responses and the code proceeds after that. These aren't
+#' used directly, but are part of how `dittodb` works.
 #'
 #' @param conn a database connection (for dispatch with these methods, it should
 #'   be of class `DBIMockConnection`)
