+++ man/mock-db-methods.Rd
@@ -96,5 +96,6 @@ class \code{DBIMockResult})}
 \description{
 Various methods (\code{dbSendQuery}, \code{dbFetchQuery}) that are mocks of the DBI
 methods of the same name. Instead of actually interacting with a database,
-they read in mock responses and the code proceeds after that.
+they read in mock responses and the code proceeds after that. These aren't
+used directly, but are part of how \code{dittodb} works.
 }
