+++ NAMESPACE
@@ -14,6 +14,7 @@ export(nycflights13_sql)
 export(nycflights13_sqlite)
 export(redact_columns)
 export(sanitize_table_id)
+export(set_dittodb_debug_level)
 export(start_db_capturing)
 export(start_mock_db)
 export(stop_db_capturing)
