### 2020-05-06 53f1a31
Add docs about debugging, some other clean up (#89)
* Add docs about debugging, some other clean up
