+++ R/utils.R
@@ -64,6 +64,33 @@ ignore_quotes <- function(statement) {
   return(gsub("(`|\"|')", "`", statement))
 }
 
+#' Set dittodb's debug level
+#'
+#' It can be helpful to see what's going on by increasing `dittodb`'s verbosity
+#' which will show what's going on under the hood (e.g. what queries are being
+#' requested, from where).
+#'
+#' The `level` argument is a numeric, where 0 is the default and (relatively)
+#' silent. The higher the level, the more verbose `dittodb` will be.
+#'
+#' Currently, `dittodb` only has one level of debugging (any value 1 or
+#' greater), but more might be used in the future.
+#'
+#' @param level a numeric, the level to set to (e.g. 1)
+#'
+#' @return the level, invisibly
+#' @export
+#'
+#' @examples
+#' set_dittodb_debug_level(1)
+#' set_dittodb_debug_level(0)
+set_dittodb_debug_level <- function(level) {
+  stopifnot(is.numeric(level))
+  options("dittodb.debug" = level)
+
+  return(invisible(level))
+}
+
 #' Get the dittodb debug level and evaluate if it is above a level
 #'
 #' @param level the level to test against (greater than or equal to)
