+++ R/connection.R
@@ -117,16 +117,6 @@ get_dbname <- function(dots) {
 #' @importFrom methods getMethod
 #'
 #' @export
-setMethod(
-  "dbListTables", signature("DBIMockConnection"),
-  function(conn, ...) {
-    getMethod("dbListTables", signature = conn@original_class)(conn, ...)
-  }
-)
-
-#' @rdname mock-db-methods
-#'
-#' @export
 setMethod("dbGetInfo", signature("DBIMockConnection"), function(dbObj, ...) {
   path <- make_path(dbObj@path, "conInfo", "")
   return(read_file(find_file(path)))
