+++ man/mock-db-methods.Rd
@@ -6,10 +6,10 @@
 \alias{mock-db-methods}
 \alias{DBIMockConnection-class}
 \alias{dbDisconnect,DBIMockConnection-method}
-\alias{dbListTables,DBIMockConnection-method}
 \alias{dbGetInfo,DBIMockConnection-method}
 \alias{dbWriteTable,DBIMockConnection,character,data.frame-method}
 \alias{dbRemoveTable,DBIMockConnection,character-method}
+\alias{dbListTables,DBIMockConnection-method}
 \alias{dbListFields,DBIMockConnection,character-method}
 \alias{dbListFields,DBIMockConnection,Id-method}
 \alias{dbListFields,DBIMockConnection,ANY-method}
@@ -29,8 +29,6 @@
 \title{Methods for interacting with DB mocks instead of an actual database}
 \usage{
 \S4method{dbDisconnect}{DBIMockConnection}(conn, ...)
-
-\S4method{dbListTables}{DBIMockConnection}(conn, ...)
 
 \S4method{dbGetInfo}{DBIMockConnection}(dbObj, ...)
 
