+++ man/dot-dittodb_env.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/recording.R
+% Please edit documentation in R/capture-requests.R
 \docType{data}
 \name{.dittodb_env}
 \alias{.dittodb_env}
