+++ man/get_redactor.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/recording.R
+% Please edit documentation in R/capture-requests.R
 \name{get_redactor}
 \alias{get_redactor}
 \title{Get the current redactor}
