+++ man/capture_requests.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/recording.R
+% Please edit documentation in R/capture-requests.R
 \name{capture_requests}
 \alias{capture_requests}
 \alias{start_db_capturing}
@@ -19,7 +19,7 @@ that matches an entry will be redacted with a standard value for the column
 type (e.g. characters will be replaced with "[redacted]")}
 }
 \value{
-NULL (invisibily)
+\code{NULL} (invisibily)
 }
 \description{
 When creating database fixtures, it can sometimes be helpful to see record
