### 2020-05-06 jonkeane/dittodb#94 @jonkeane
Auto no lint fixture files
Fixture files can frequently not pass listing. Should we automatically include fixture directories to no lint if it exists? Provide a better / easier functionality for making sure that isn't too painful?
