+++ R/connection.R
@@ -78,72 +78,3 @@ dbMockConnect <- function(drv, ...) {
 
   return(new(mock_class, path = path, original_class = original_class))
 }
-
-
-#' Get the `dbname` from a connection call
-#'
-#' @param dots from the argument being passed to the connection
-#'
-#' @return the name, sanitized if needed
-#'
-#' @keywords internal
-#'
-#' @export
-get_dbname <- function(dots) {
-  # look through dots to grab either dbname or the first unnammed argument
-  named_dbname <- !is.null(dots$dbname) && dots$dbname != ""
-  named_database <- !is.null(dots$Database) && dots$Database != ""
-  unnamed_dbname <- length(dots) > 0 &&
-    (is.null(names(dots[1])) || names(dots[1]) == "")
-  # if there is no name, or it's empty
-  if (named_dbname) {
-    path <- dots$dbname
-  } else if (named_database) {
-    path <- dots$Database
-  } else if (unnamed_dbname) {
-    path <- dots[[1]]
-  } else {
-    stop("There was no dbname, so I don't know where to look for mocks.")
-  }
-  return(db_path_sanitize(path))
-}
-
-
-# DBI methods from the respective DBI DB implementations
-# TODO: make a list of these so that it's not just copy/paste?
-
-#' @rdname mock-db-methods
-#'
-#' @importFrom methods getMethod
-#'
-#' @export
-setMethod(
-  "dbListTables", signature("DBIMockConnection"),
-  function(conn, ...) {
-    getMethod("dbListTables", signature = conn@original_class)(conn, ...)
-  }
-)
-
-#' @rdname mock-db-methods
-#'
-#' @export
-setMethod("dbGetInfo", signature("DBIMockConnection"), function(dbObj, ...) {
-  path <- make_path(dbObj@path, "conInfo", "")
-  return(read_file(find_file(path)))
-})
-
-# TODO: should we also implement signature("DBIMockConnection", "character",
-# "character") most backends warn that it isn't well supported.
-#' @rdname mock-db-methods
-#' @export
-setMethod(
-  "dbWriteTable", signature("DBIMockConnection", "character", "data.frame"),
-  function(conn, name, value, ...) return(TRUE)
-)
-
-#' @rdname mock-db-methods
-#' @export
-setMethod(
-  "dbRemoveTable", signature("DBIMockConnection", "character"),
-  function(conn, name, ...) return(TRUE)
-)
