### 2020-05-06 20a9300
capture_db_requests
