### 2020-05-07 c814576
Capture db requests (#95)
