+++ inst/nycflight_mocks/nycflights/SELECT-ef6317.R
@@ -1,2 +1,2 @@
-structure(list(carrier = "9E", name = "Endeavor Air Inc."), class = "data.frame", row.names = c(NA, 
+structure(list(carrier = "9E", name = "Endeavor Air Inc."), class = "data.frame", row.names = c(NA,
 -1L))
