+++ presentations/review_response.md
@@ -24,7 +24,8 @@ Community guidelines comments:
 Other comments:
 * Starting and stopping mocks. Both reviewers mentioned wanting an additional way to start and stop mocking besides `with_mock_db()`. I've added `start_mock_db()` and `stop_mock_db()` to turn mocking on and off. This way people can step through their tests if they want (and this addresses some of the other comments as well). 
 
-* Organization *TODO*
+
+* Organization I've reorganized some of the methods to be more thematic and all of the methods that work with the various `DBI` methods (e.g. `DBI::dbQueries-Results.R` for query and result focused methods).
 
 * Debugging: thank you for the suggestion, I've added `set_dittodb_debug_level()` and included some documentation about how to use it and what is printed with debugging turned on.
 
@@ -35,4 +36,110 @@ Other comments:
 * Developer setup, I've added a whole vignette devoted to how to develop `dittodb` and to set up testing databases. I've also reorganized some of the setup scripts to make it easier to see which ones to use, how to turn on the tests for those when you're ready. Finally, I've added a method for changing the ports that are used by the test databases. 
 
 
+@etiennebr's comments
 
+Installation instructions:
+* I've added a new givnette that describes the process to set up development databases, various configuration options (when, why, and how to turn on or off tests locally). I've also clarified and described more about what the databse setup scripts are for (and warned against running ones that are intended for / needed for CI compared to ones that are intended for running locally)
+
+Improved vignettes
+*TODO*
+>  I would suggest that a vignette walking the user through the
+full process of testing during package development would make adoption much
+easier. I'll comment further on this later, but I could not figure a smooth
+workflow, which I'm sure the authors have figured out.
+> 
+> I find the diversity of examples in the vignettes and the README very
+relevant, but as I mentionned above, database setup could be improved to
+accelerate the users. I would suggest to:
+> 
+> Provide a simple example using maybe a simple sqlite file in the readme;
+Provide instructions to run a docker container in the vignettes.
+
+Community Guidlines:
+* I've added a contributing guidlines document, thanks for noticing this oversight.
+
+Automated Tests:
+* I've added more description of what tests are not run by defualt locally (which makes coverage lower like you noted). To get full coverage one has to run all of the database backends during tests and turn on those tests with environment variables (both to make local development easier when one is not messing with some backends as well as to make sure these tests are not run on CRAN). The CI setup turns these on (and sets up the databses) and the coverage figure comes from there.
+
+* Thanks for noticing I wasn't recommending the correct `.sh` scripts in my very brief discussion in the readme. I've expanded that discussion in the new development vignette and I renamed some of those scripts to hopefully be more user-friendly and obvious.
+
+
+Workflow:
+*TODO* 
+> As I mentioned earlier, I could understand the vignettes and the examples, but
+> I'm not sure I understand dittodb intented package development workflow. By
+> looking at the package's own tests, I think I could grapple some ideas, but I
+> still find the ergonomic of the package could be better (or explained). Maybe I
+> should just get more familiar with it, but I feel like I spent a larger amount
+> of time playing with it than most potential users would --out of pure
+> motivation, mind you.
+> 
+> Here's how I used dittodb:
+> 
+> add a start and stop line at the beginning and the end of the database tests
+> run the tests manually
+> remove the start and stop line
+> wrap all the code in a with_mock_db({})
+> repeat every time you add a test (and make sure contributors know the gymnastic)
+> Here are my pain points:
+> 
+> Debugging is harder, because I can't step through the tests
+> We need to change many lines of code in different places every time a test
+> changes, which makes it error prone
+> The commit that adds dittodb change all the lines in a test file
+> (re-indent), which burries the changes in another layer (maybe that's just me)
+> What I would like is to:
+> 
[…]
