+++ man/get_dbname.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/connection.R
+% Please edit documentation in R/utils.R
 \name{get_dbname}
 \alias{get_dbname}
 \title{Get the \code{dbname} from a connection call}
