+++ man/mock-db-methods.Rd
@@ -1,20 +1,15 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/connection.R, R/listTables.R, R/quote.R,
-%   R/result.R
+% Please edit documentation in R/connection.R, R/dbListTables-Fields.R,
+%   R/dbQueries-Results.R, R/dbMisc.R, R/quote.R
 \docType{class}
 \name{mock-db-methods}
 \alias{mock-db-methods}
 \alias{DBIMockConnection-class}
 \alias{dbDisconnect,DBIMockConnection-method}
-\alias{dbGetInfo,DBIMockConnection-method}
-\alias{dbWriteTable,DBIMockConnection,character,data.frame-method}
-\alias{dbRemoveTable,DBIMockConnection,character-method}
 \alias{dbListTables,DBIMockConnection-method}
 \alias{dbListFields,DBIMockConnection,character-method}
 \alias{dbListFields,DBIMockConnection,Id-method}
 \alias{dbListFields,DBIMockConnection,ANY-method}
-\alias{dbQuoteIdentifier,DBIMockRPostgresConnection,character-method}
-\alias{dbQuoteIdentifier,DBIMockRPostgresConnection,SQL-method}
 \alias{DBIMockResult-class}
 \alias{dbSendQuery,DBIMockConnection,character-method}
 \alias{dbFetch,DBIMockResult-method}
@@ -24,17 +19,16 @@
 \alias{dbClearResult,DBIMockResult-method}
 \alias{dbHasCompleted,DBIMockResult-method}
 \alias{dbGetQuery,DBIMockRPostgreSQLConnection,character-method}
+\alias{dbGetInfo,DBIMockConnection-method}
+\alias{dbWriteTable,DBIMockConnection,character,data.frame-method}
+\alias{dbRemoveTable,DBIMockConnection,character-method}
 \alias{dbColumnInfo,DBIMockResult-method}
 \alias{dbGetInfo,DBIMockResult-method}
+\alias{dbQuoteIdentifier,DBIMockRPostgresConnection,character-method}
+\alias{dbQuoteIdentifier,DBIMockRPostgresConnection,SQL-method}
 \title{Methods for interacting with DB mocks instead of an actual database}
 \usage{
 \S4method{dbDisconnect}{DBIMockConnection}(conn, ...)
-
-\S4method{dbGetInfo}{DBIMockConnection}(dbObj, ...)
-
-\S4method{dbWriteTable}{DBIMockConnection,character,data.frame}(conn, name, value, ...)
-
-\S4method{dbRemoveTable}{DBIMockConnection,character}(conn, name, ...)
 
 \S4method{dbListTables}{DBIMockConnection}(conn, ...)
 
@@ -43,10 +37,6 @@
 \S4method{dbListFields}{DBIMockConnection,Id}(conn, name, ...)
 
 \S4method{dbListFields}{DBIMockConnection,ANY}(conn, name, ...)
-
-\S4method{dbQuoteIdentifier}{DBIMockRPostgresConnection,character}(conn, x, ...)
-
-\S4method{dbQuoteIdentifier}{DBIMockRPostgresConnection,SQL}(conn, x, ...)
 
 \S4method{dbSendQuery}{DBIMockConnection,character}(conn, statement, ...)
 
@@ -64,9 +54,19 @@
 
 \S4method{dbGetQuery}{DBIMockRPostgreSQLConnection,character}(conn, statement, ...)
 
+\S4method{dbGetInfo}{DBIMockConnection}(dbObj, ...)
+
+\S4method{dbWriteTable}{DBIMockConnection,character,data.frame}(conn, name, value, ...)
+
+\S4method{dbRemoveTable}{DBIMockConnection,character}(conn, name, ...)
[…]

[…]
