### 2020-05-11 bba2ad1
Use a different dbplyr set of verbs to avoid a lifecycle deprecation in progress
