+++ vignettes/travelling/resultInfo-6f7e1a.R
@@ -1,0 +1,18 @@
+list(statement = "SELECT * FROM \"flights\" AS \"zzz13\" WHERE 0=1", 
+    isSelect = 1L, rowsAffected = -1L, rowCount = 0L, completed = 0L, 
+    fieldDescription = list(list(name = c("year", "month", "day", 
+    "dep_time", "sched_dep_time", "dep_delay", "arr_time", "sched_arr_time", 
+    "arr_delay", "carrier", "flight", "tailnum", "origin", "dest", 
+    "air_time", "distance", "hour", "minute", "time_hour"), Sclass = c(13L, 
+    13L, 13L, 13L, 13L, 14L, 13L, 13L, 14L, 16L, 13L, 16L, 16L, 
+    16L, 14L, 14L, 14L, 14L, 16L), type = c(23L, 23L, 23L, 23L, 
+    23L, 701L, 23L, 23L, 701L, 25L, 23L, 25L, 25L, 25L, 701L, 
+    701L, 701L, 701L, 1184L), len = c(4L, 4L, 4L, 4L, 4L, 8L, 
+    4L, 4L, 8L, -1L, 4L, -1L, -1L, -1L, 8L, 8L, 8L, 8L, 8L), 
+        precision = c(-1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, 
+        -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L
+        ), scale = c(-1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, 
+        -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L, -1L
+        ), nullOK = c(TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, 
+        TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, 
+        TRUE, TRUE, TRUE))))
