+++ db-setup/postgres-brew.sh
@@ -1,7 +1,8 @@
 rm -rf /usr/local/var/postgres
 # for odbc tests, we need to also install the odbc drivers
-# we might also need to install unixodbc here if it's not already installed.
-brew install postgresql psqlodbc
+# we might also need to install unixodbc here if it's not already installed
+# we might also need to install postgresql here if it's not already installed
+brew install psqlodbc
 cat <<EOT >> /usr/local/etc/odbcinst.ini
 [PostgreSQL Unicode]
 Description     = PostgreSQL ODBC driver (Unicode 9.2)
