### 2020-05-11 1431336
Remove group by (#97)
Use a different dbplyr set of verbs to avoid a lifecycle deprecation in progress
