+++ vignettes/travelling/dittodb/SELECT-dfae22.R
@@ -1,308 +1,0 @@
-structure(list(tailnum = c("N4XKAA", "N904WN", "N3BWAA", "N3CJAA",
-                           "N67052", "N521JB", "N8307K", "N662AW", "N8974C", "N8855A", "N409WN",
-                           "N640MQ", "N173DZ", "N394AA", "N926AT", "N11535", "N505AA", "N4XYAA",
-                           "N393DA", "N639VA", "N34222", "N77518", "N511MQ", "N651JB", "N34131",
-                           "N643JB", "N988DL", "N792SW", "N231JB", "N676DL", "N14542", "N531JB",
-                           "N14573", "N13161", "N201LV", "N78506", "N27962", "N287WN", "N491AA",
-                           "N198JB", "N8432A", "N14902", "N8EGMQ", "N569AA", "N3FJAA", "N336NB",
-                           "N905DL", "N628VA", "N12136", "N550WN", "N844VA", "N738US", "N371NB",
-                           "N431WN", "N11206", "N412WN", "N154UW", "N832UA", "N14993", "N325NB",
-                           "N17122", "N3759", "N314US", "N14231", "N534UW", "N196UW", "N638MQ",
-                           "N242WN", "N764US", "N802MQ", "N33209", "N21108", "N421UA", "N455UW",
-                           "N3FCAA", "N672DL", "N682MQ", "N173US", "N756SA", "N691CA", "N945WN",
-                           "N730US", "N479UA", "N180US", "N903FJ", "N513UA", "N775JB", "N367AA",
-                           "N677UA", "N955DL", "N611QX", "N7732A", "N11192", "N36280", "N338AA",
-                           "N11113", "N14106", "N3CGAA", "N477AA", "N837UA", "N618DL", "N364AA",
-                           "N381HA", "N8869B", "N649DL", "N284JB", "N924DL", "N580AA", "N3744F",
-                           "N416WN", "N14242", "N7702A", "N324JB", "N833UA", "N704X", "N506UA",
-                           "N722US", "N14904", "N764NC", "N816UA", "N591AA", "N965AT", "N539MQ",
-                           "N768JB", "N358NB", "N554UA", "N481UA", "N5DAAA", "N444UA", "N69806",
-                           "N292PQ", "N413AS", "N75410", "N13138", "N368AA", "N3AUAA", "N299PQ",
-                           "N306JB", "N710UW", "N76064", "N971AT", "N617MQ", "N8891A", "N33266",
-                           "N823MQ", "N8896A", "N398DA", "N944WN", "N915DE", "N3HTAA", "N592UA",
-                           "N193DN", "N585UA", "N939AT", "N4XRAA", "N265WN", "N4YFAA", "N281WN",
-                           "N838VA", "N785SW", "N267AT", "N4WAAA", "N523UA", "N912DL", "N27213",
-                           "N707TW", "N844UA", "N490WN", "N17133", "N386HA", "N11119", "N608QX",
-                           "N534JB", "N76526", "N8311Q", "N8921B", "N18102", "N472AA", "N7714B",
-                           "N535MQ", "N529AA", "N8982A", "N3DWAA", "N947WN", "N631AA", "N735MQ",
-                           "N909WN", "N13994", "N348AA", "N931MQ", "N948WN", "N722MQ", "N282WN",
-                           "N502UA", "N13202", "N26226", "N74856", "N404WN", "N842VA", "N967UW",
[…]
