+++ vignettes/travelling.Rmd
@@ -13,36 +13,44 @@ vignette: >
 library(dplyr)
 library(dbplyr)
 
-con_psql <- RPostgreSQL::dbConnect(
+con_psql <- DBI::dbConnect(
   drv = DBI::dbDriver("PostgreSQL"),
-  dbname = "dittodb",
-  host = "postgres.server",
-  user = "dittodb",
-  password = "_dittodb_"
+  dbname = "",
+  host = "127.0.0.1",
+  user = getOption("dittodb.test.user"),
+  password = getOption("dittodb.test.pw")
 )
+DBI::dbSendStatement(con, "CREATE DATABASE travelling")
+DBI::dbDisconnect(con_psql)
 
+
+con_psql <- DBI::dbConnect(
+  drv = DBI::dbDriver("PostgreSQL"),
+  dbname = "travelling",
+  host = "127.0.0.1",
+  user = getOption("dittodb.test.user"),
+  password = getOption("dittodb.test.pw")
+)
 nycflights13_sql(con_psql)
-
-dbDisconnect(con_psql)
+DBI::dbDisconnect(con_psql)
 
 start_db_capturing(path = "vignettes/")
-
-con_psql <- RPostgreSQL::dbConnect(
+con_psql <- DBI::dbConnect(
   drv = DBI::dbDriver("PostgreSQL"),
-  dbname = "dittodb",
-  host = "postgres.server",
-  user = "dittodb",
-  password = "_dittodb_"
+  dbname = "travelling",
+  host = "127.0.0.1",
+  user = getOption("dittodb.test.user"),
+  password = getOption("dittodb.test.pw")
 )
 
-tbl(con_psql, in_schema("public", "flights")) %>%
+tbl(con_psql, "flights") %>%
   filter(!is.na(tailnum)) %>%
-  group_by(tailnum) %>%
-  count() %>%
-  filter(n >= 10L)
+  filter(arr_delay >= 180) %>% 
+  select(tailnum) %>%
+  distinct() %>% 
+  collect()
 
 dbDisconnect(con_psql)
-
 stop_db_capturing()
 ```
 
@@ -73,11 +81,12 @@ possible to run outside the office (or even with no internet access at all!).
 # Recording queries
 
 Suppose we are asked to analyze the flights to only show flights with planes 
-that have flown at least 100 flights.
+that have been delayed at least 3 hours.
 
-One would find all planes that have flown at least 100 flights. The only 
-consideration would be to filter those flights with missing tail number or those
-will be treated as a single plane.
+One would find all the flights that have been delayed by over 3 hours, and then 
+only grab the distinct tail numbers. The only consideration would be to filter 
+those flights with missing tail number or those will be treated as a single 
+plane.
 
 We could run the following code to get that data with a direct connection to the 
 database (i.e. at the office):
@@ -85,19 +94,18 @@ database (i.e. at the office):
 library(dplyr)
 library(dbplyr)
 
-con_psql <- RPostgreSQL::dbConnect(
+con_psql <- DBI::dbConnect(
   drv = DBI::dbDriver("PostgreSQL"),
-  dbname = "dittodb",
-  host = "postgres.server",
-  user = "dittodb",
-  password = "_dittodb_"
+  dbname = "travelling",
+  host = "127.0.0.1",
+  user = "m.ciccone"
 )
 
-tbl(con_psql, in_schema("public", "flights")) %>%
+tbl(con_psql, "flights") %>%
   filter(!is.na(tailnum)) %>%
-  group_by(tailnum) %>%
-  count() %>%
-  filter(n >= 10L)
+  filter(arr_delay >= 180) %>% 
+  select(tailnum) %>%
+  distinct()
 ```
 
[…]
