### 2020-05-26 a08349f
catch even not-running docker containers when restarting
