+++ db-setup/local-postgres-docker-setup.sh
@@ -1,4 +1,4 @@
-if [ -z "$(docker ps -q -f name=dittodb-postgres)" ]
+if [ -z "$(docker ps -a -q -f name=dittodb-postgres)" ]
 then
   echo "Starting docker container fresh..."
 else
