+++ presentations/satRday2020/satRday2020.Rmd
@@ -1,0 +1,384 @@
+---
+title: "dittodb"
+subtitle: "making testing databases easy"
+author: "Jonathan Keane"
+institute: "Socure"
+date: "2020-05-30 (updated: `r Sys.Date()`)"
+output:
+  xaringan::moon_reader:
+    lib_dir: libs
+    css: ["xaringan-themer.css", "Azuro.css"]
+    nature:
+      highlightStyle: github
+      highlightLines: true
+      countIncrementalSlides: false
+      slideNumberFormat: ""
+---
+background-image: url("images/IMG_5793.png")
+background-position: 50% 50%
+background-size: cover
+
+class: inverse, center, top
+
+
+# What is testing?
+
+```{r setup, include=FALSE}
+options(htmltools.dir.version = FALSE)
+```
+
+```{r xaringan-themer, include=FALSE, warning=FALSE}
+library(emojifont) 
+library(xaringanthemer)
+style_mono_accent(
+  base_color = "#1c5253",
+  header_font_family = "Azuro",
+  text_font_family = "Azuro",
+  code_font_google   = google_font("Fira Mono"),
+  text_font_size = "1.25rem",
+  colors = c(white = "#FFFFFF", gray = "#898C9C"),
+  header_h2_font_size = "1.75rem",
+  header_h3_font_size = "1.25rem",
+  extra_css = list(
+    ".large" = list(
+      "font-size" = "2.75rem"
+    ),
+    ".inverse > h1 > code" = list(
+      "color" = "var(--inverse-header-color)"
+    ),
+    ".colophon" = list(
+      "font-size" = "0.9rem",
+      "position" = "absolute",
+      "bottom" = 0
+    ),
+    ".flex-box" = list(
+    "display" = "flex",
+    "flex-direction" = "column",
+    "justify-content" = "space-between",
+    "height" = "70%"
+    ),
+    # the first p of .flext-box should have no top-margin
+    ".flex-box > p:first-child" = list(
+      "margin-top" = 0
+    ),
+    ".flex-box > p:last-of-type" = list(
+      "padding-bottom" = "1em"
+    ),
+    ".seventyfive" = list(
+      "width" = "75%",
+      "margin" = "0 auto"
+    )
+  )
+)
+```
+
+---
+class: inverse, center, middle
+
+## A way to check that the code you wrote is doing what you expect it to.
+
+---
+# Easier said than done
+
+<br /><br />
+Any test is better than no tests
+
+--
+<br /><br />
+
+but a good test is worth its (metaphorical) weight in gold
+
+---
+# A good test is ...
+* quick to run
+* expressive
+* doesn't require special set up
+
+--
+
+# But databases are ...
+* (relatively) slow
+* rely on external state or set up
+* requires a connection or custom setup
+
+---
+# CI + DBs = ☠️
+
+Continuous integration (like GitHub Actions, Travis, AppVeyor) is wonderful, 
+you can run your code on someone else's ☁️ 💻 
+<br/> <br/>
+
+--
+
+But then you're running your code in an environment you have to set up.
+--
+ That you can't interactively iterate on.
+<br/> <br/>
+
+--
+
+Even installing R dependencies can be a pain.
+
+---
+background-image: url("images/db-test-ci-db-hard-more-again.png")
+background-position: 50% 50%
+background-size: contain
+class: center, bottom
+
+---
+background-image: url("images/db-test-ci-db-hard-more.png")
+background-position: 0% 20%
+background-size: contain
+class: center, bottom
+
[…]
