+++ presentations/satRday2020/satRday2020.Rmd
@@ -1,8 +1,8 @@
 ---
 title: "dittodb"
 subtitle: "making testing databases easy"
-author: "Jonathan Keane"
-institute: "Socure"
+author: "Jonathan Keane · Socure"
+institute: "Chicago satRdays 2020"
 date: "2020-05-30 (updated: `r Sys.Date()`)"
 output:
   xaringan::moon_reader:
@@ -37,8 +37,6 @@ style_mono_accent(
   code_font_google   = google_font("Fira Mono"),
   text_font_size = "1.25rem",
   colors = c(white = "#FFFFFF", gray = "#898C9C"),
-  header_h2_font_size = "1.75rem",
-  header_h3_font_size = "1.25rem",
   extra_css = list(
     ".large" = list(
       "font-size" = "2.75rem"
@@ -67,6 +65,25 @@ style_mono_accent(
     ".seventyfive" = list(
       "width" = "75%",
       "margin" = "0 auto"
+    ),
+    ".title-slide h1" = list(
+      "font-size" = "4rem",
+      "margin-bottom" = "0rem"
+      ),
+    ".title-slide h2" = list(
+      "font-size" = "1.75rem",
+      "margin-top" = "0rem",
+      "margin-bottom" = "4rem"
+      ),
+    ".title-slide h3" = list(
+      "font-size" = "1.75rem",
+      "margin-block-start" = "0rem",
+      "margin-block-end" = "0rem"
+      ),
+    ".title-slide > h3:last-of-type" = list(
+      "font-size" = "1rem",
+      "margin-block-start" = "4rem",
+      "margin-block-end" = "2rem"
     )
   )
 )
@@ -189,7 +206,10 @@ On CI we “just” do the same:
 ---
 # `dittodb` lets you ...
 
-Write tests that act like they are connected to a real-live database using `with_mock_db({...})`
+Write tests that act like they are connected to a real-live database using 
+`with_mock_db({...})` without ever having to set one up.
+
+<br/ >
 
 ```r
 with_mock_db({
@@ -247,10 +267,6 @@ Much of the above is from https://www.martinfowler.com/articles/mocksArentStubs.
 * `r emoji("scream")`
 ]
 
-???
-
-If <5 minutes type 26 to go to the next section
-
 ---
 background-image: url("images/_DSC1098.png")
 background-position: 50% 50%
@@ -259,55 +275,9 @@ class: center, top
 
 # Demo
 
----
-# Mandatory plug
-.flex-box[
-## Socure is hiring — [socure.com/about/careers](https://www.socure.com/about/careers)
+???
 
-.center.seventyfive[![](https://www.socure.com/hubfs/Assets_2018/SOC-logo-color.svg)]
-]
-
----
-
-# Thanks!
-.flex-box[
-`dittodb` lives at [_dittodb.jonkeane.com_](https://dittodb.jonkeane.com)
-
-Install `dittodb` with `remotes::install_github("jonkeane/dittodb")`
-
-Extra special thanks to Mauricio Vargas @pachamaltese and thanks to [`httptest`](https://enpiar.com/r/httptest) for inspiration.
-
-@jonkeane 
-&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-[_jonkeane.com_](https://jonkeane.com) 
[…]
