### 2020-05-30 85c679e
Cleanups from presentations branch (#105)
Move presentations to a new repo
