+++ R/use-dittodb.R
@@ -1,0 +1,81 @@
+#' Use 'dittodb' in your tests
+#'
+#' This function adds `dittodb` to Suggests in the package DESCRIPTION and
+#' loads it in `tests/testthat/helpers.R`. Call it once when you're setting up
+#' a new package test suite.
+#'
+#' The function is idempotent: if `dittodb` is already added to these files, no
+#' additional changes will be made.
+#'
+#' @param path character path to the package
+#' @importFrom utils tail
+#' @return Nothing: called for file system side effects.
+#' @export
+use_dittodb <- function (path=".") {
+  if (!("DESCRIPTION" %in% dir(path))) {
+    stop(path, " is not an R package directory", call.=FALSE)
+  }
+  add_dittodb_to_desc(file.path(path, "DESCRIPTION"))
+  # TODO: could allow helper.r too
+  add_dittodb_to_helper(file.path(path, "tests", "testthat", "helpers.R"))
+  invisible()
+}
+
+add_dittodb_to_desc <- function (file) {
+  # Read DESCRIPTION, add dittodb to Suggests if not already there
+
+  # Hack to preserve whitespace: read it twice
+  desc_fields <- colnames(read.dcf(file))
+  desc <- read.dcf(file, keep.white=desc_fields)
+  if (!("Suggests" %in% desc_fields)) {
+    # Add a column for Suggests
+    desc <- cbind(desc, matrix("", ncol=1, dimnames=list(NULL, "Suggests")))
+  }
+  if (!grepl("dittodb", desc[,"Suggests"])) {
+    # Add dittodb
+
+    # Parse the list, and try to preserve the whitespace from the original
+    suggested_pkgs <- unlist(strsplit(desc[,"Suggests"], ","))
+    sep <- sub("^([[:blank:]\n]*).*", "\\1", suggested_pkgs)
+    suggested_pkgs <- sort(c(trimws(suggested_pkgs), "dittodb"))
+    extra_sep <- tail(unique(sep), 1)
+    if (length(extra_sep) == 0 || nchar(extra_sep) == 0) {
+      extra_sep <- " "
+    }
+    sep <- c(sep, extra_sep)
+    desc[,"Suggests"] <- paste0(sep, suggested_pkgs, collapse=",")
+
+    # Msg and write
+    message("Adding 'dittodb' to Suggests in DESCRIPTION")
+    write.dcf(desc, file=file, keep.white=desc_fields)
+  }
+}
+
+add_dittodb_to_helper <- function (file) {
+  # Create tests/testthat/helpers.R if it does not exist
+
+  if (!file.exists(file)) {
+    message("Creating ", file)
+    message("Adding library(dittodb) to ", file)
+    mkdir_p(file)
+    cat("library(dittodb)\n", file=file)
+    # Msg and write
+  } else {
+    helper_lines <- readLines(file)
+    if (!any(grepl("library(dittodb)", helper_lines, fixed=TRUE))) {
+      # Add "library(dittodb)" to the top if it's not already there
+      helper_lines <- c("library(dittodb)", helper_lines)
+      # Msg and write
+      message("Adding library(dittodb) to ", file)
+      writeLines(helper_lines, file)
+    }
+  }
+}
+
+mkdir_p <- function (filename) {
+  # Recursively create the directories so that we can write this file.
+  # If they already exist, do nothing.
+  # Like mkdir -p path
+
+  dir.create(dirname(filename), showWarnings=FALSE, recursive=TRUE)
+}
