+++ man/use_dittodb.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/use-dittodb.R
+\name{use_dittodb}
+\alias{use_dittodb}
+\title{Use 'dittodb' in your tests}
+\usage{
+use_dittodb(path = ".")
+}
+\arguments{
+\item{path}{character path to the package}
+}
+\value{
+Nothing: called for file system side effects.
+}
+\description{
+This function adds \code{dittodb} to Suggests in the package DESCRIPTION and
+loads it in \code{tests/testthat/helpers.R}. Call it once when you're setting up
+a new package test suite.
+}
+\details{
+The function is idempotent: if \code{dittodb} is already added to these files, no
+additional changes will be made.
+}
