+++ NAMESPACE
@@ -20,6 +20,7 @@ export(start_db_capturing)
 export(start_mock_db)
 export(stop_db_capturing)
 export(stop_mock_db)
+export(use_dittodb)
 export(with_mock_db)
 export(with_mock_path)
 exportClasses(DBIMockConnection)
@@ -54,4 +55,5 @@ importFrom(methods,setMethod)
 importFrom(testthat,with_mock)
 importFrom(utils,head)
 importFrom(utils,installed.packages)
+importFrom(utils,tail)
 useDynLib(dittodb, .registration = TRUE)
