### 2020-05-30 d198011
ignore use-ditto.R for coverage
