+++ codecov.yml
@@ -10,3 +10,7 @@ coverage:
       default:
         target: auto
         threshold: 1%
+
+coverage:
+  ignore:
+    - "R/use-dittodb.R"  # this internal function is called just for testing
