### 2020-05-31 1165a04
Tests for use_dittodb, renaming helpers.R to the more standard helper.R (to match helper-*.R), an da bit of clean up
