### 2020-05-31 151b987
Clean up linting
