+++ R/use-dittodb.R
@@ -11,9 +11,9 @@
 #' @importFrom utils tail
 #' @return Nothing: called for file system side effects.
 #' @export
-use_dittodb <- function (path=".") {
+use_dittodb <- function(path = ".") {
   if (!("DESCRIPTION" %in% dir(path))) {
-    stop(path, " is not an R package directory", call.=FALSE)
+    stop(path, " is not an R package directory", call. = FALSE)
   }
   add_dittodb_to_desc(file.path(path, "DESCRIPTION"))
   # TODO: could allow helper.r too
@@ -21,48 +21,48 @@ use_dittodb <- function (path=".") {
   invisible()
 }
 
-add_dittodb_to_desc <- function (file) {
+add_dittodb_to_desc <- function(file) {
   # Read DESCRIPTION, add dittodb to Suggests if not already there
 
   # Hack to preserve whitespace: read it twice
   desc_fields <- colnames(read.dcf(file))
-  desc <- read.dcf(file, keep.white=desc_fields)
+  desc <- read.dcf(file, keep.white = desc_fields)
   if (!("Suggests" %in% desc_fields)) {
     # Add a column for Suggests
-    desc <- cbind(desc, matrix("", ncol=1, dimnames=list(NULL, "Suggests")))
+    desc <- cbind(desc, matrix("", ncol = 1, dimnames = list(NULL, "Suggests")))
   }
-  if (!grepl("dittodb", desc[,"Suggests"])) {
+  if (!grepl("dittodb", desc[, "Suggests"])) {
     # Add dittodb
 
     # Parse the list, and try to preserve the whitespace from the original
-    suggested_pkgs <- unlist(strsplit(desc[,"Suggests"], ","))
+    suggested_pkgs <- unlist(strsplit(desc[, "Suggests"], ","))
     sep <- sub("^([[:blank:]\n]*).*", "\\1", suggested_pkgs)
     suggested_pkgs <- sort(c(trimws(suggested_pkgs), "dittodb"))
     extra_sep <- tail(unique(sep), 1)
-    if (length(extra_sep) == 0 || nchar(extra_sep) == 0) {
+    if (length(extra_sep)  == 0 || nchar(extra_sep) == 0) {
       extra_sep <- " "
     }
     sep <- c(sep, extra_sep)
-    desc[,"Suggests"] <- paste0(sep, suggested_pkgs, collapse=",")
+    desc[, "Suggests"] <- paste0(sep, suggested_pkgs, collapse = ",")
 
     # Msg and write
     message("Adding 'dittodb' to Suggests in DESCRIPTION")
-    write.dcf(desc, file=file, keep.white=desc_fields)
+    write.dcf(desc, file = file, keep.white = desc_fields)
   }
 }
 
-add_dittodb_to_helper <- function (file) {
+add_dittodb_to_helper <- function(file) {
   # Create tests/testthat/helper.R if it does not exist
 
   if (!file.exists(file)) {
     message("Creating ", file)
     message("Adding library(dittodb) to ", file)
-    dir.create(dirname(file), showWarnings=FALSE, recursive=TRUE)
-    cat("library(dittodb)\n", file=file)
+    dir.create(dirname(file), showWarnings = FALSE, recursive = TRUE)
+    cat("library(dittodb)\n", file = file)
     # Msg and write
   } else {
     helper_lines <- readLines(file)
-    if (!any(grepl("library(dittodb)", helper_lines, fixed=TRUE))) {
+    if (!any(grepl("library(dittodb)", helper_lines, fixed = TRUE))) {
       # Add "library(dittodb)" to the top if it's not already there
       helper_lines <- c("library(dittodb)", helper_lines)
       # Msg and write
