+++ .github/workflows/check-linux-ubuntu.yaml
@@ -31,6 +31,9 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
 
+      - name: test
+        run: Rscript -e 'installed.packages()["tools",]'
+
       - uses: r-lib/actions/setup-pandoc@master
 
       - uses: r-lib/actions/setup-tinytex@master
