+++ .github/workflows/check-linux-ubuntu.yaml
@@ -84,7 +84,7 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: test R cmd
-        run: R CMD build
+        run: R CMD build .
 
       - name: Check
         run: Rscript -e "rcmdcheck::rcmdcheck(args = '${{ matrix.config.args }}', error_on = 'note', check_dir = 'check')"
