### 2020-06-03 dc6f291
Install dev rmarkdown for rstudio/rmarkdown#1832
