### 2020-06-03 380180c
dev rmarkdown for vignettes <3.6 (#108)
Install dev rmarkdown for rstudio/rmarkdown#1832
