+++ .github/workflows/check-linux-ubuntu.yaml
@@ -70,6 +70,9 @@ jobs:
       - name: Install R dependencies
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'httptest'), type = 'source')" -e "remotes::install_deps(dependencies = TRUE);"
 
+      - name: Install dev rmarkdown
+        run: Rscript -e "remotes::install_github('rstudio/rmarkdown')"
+
       - name: Build R DB Dependencies on cache miss
         if: steps.rlibcache.outputs.cache-hit != 'true'
         run: Rscript -e "install.packages(c('RMariaDB', 'odbc', 'RPostgreSQL', 'RPostgres'), type = 'source');"
