### 2020-06-03 5e5bb3f
add `use_dittodb()` to the readme+dittodb vignette
