### 2020-06-03 bd98d75
Reduce the cyclomatic complexity of request capturing (#98)
* Some shuffling
* oldrel binary fix
