### 2020-06-03 2078884
Clean up so that pkgdown is happy
