+++ vignettes/developing-dittodb.Rmd
@@ -4,7 +4,7 @@ author: "Jonathan Keane"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{developing-dittodb}
+  %\VignetteIndexEntry{Developing `dittodb`}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
