+++ vignettes/dittodb.Rmd
@@ -9,6 +9,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
+
 ```{r, include = FALSE}
 library(testthat)
 library(dittodb)
@@ -22,8 +23,6 @@ db_mock_paths("dittodb")
 ```
 
 `dittodb` is designed to make it easy and fun to test functions that interact with a database. It works by looking for mock responses for each query you send while you run your tests and will seamlessly pretend that those mocks were provided by the database connection without needing a connection at all.
-
-To get started, imagine that we are working on a package that queries a database that consists of the [nycflights13 data](nycflights.html). We have the following function which takes a column to aggregate by and returns a dataframe with that column and the mean delay for groups based on the values in the column name given.
 
 ```{r setup db, include = FALSE, eval = FALSE}
 # setup the DB used in the rest of the vignette
@@ -46,17 +45,17 @@ out <- mean_delays("month")
 stop_db_capturing()
 ```
 
+# Testing `mean_delays()` {.tabset}
+To get started, imagine that we are working on a package that queries a database that consists of the [nycflights13 data](nycflights.html). We have the following function which takes a column to aggregate by and returns a dataframe with that column and the mean delay for groups based on the values in the column name given.
+
+## RMariaDB
 ```{r mean_delays}
 library(DBI)
-library(RMariaDB)
 
 mean_delays <- function(group_col) {
   con <- dbConnect(
     RMariaDB::MariaDB(),
-    dbname = "nycflights",
-    host = "127.0.0.1",
-    username = "travis",
-    password = ""
+    dbname = "nycflights"
   )
   on.exit(dbDisconnect(con))
   
@@ -68,6 +67,45 @@ mean_delays <- function(group_col) {
   return(dbGetQuery(con, query))
 }
 ```
+
+## RPostgres
+```{r mean_delays_rpostgres, eval = FALSE}
+library(DBI)
+
+mean_delays <- function(group_col) {
+  con <- dbConnect(
+    RPostgres::Postgres(),
+    dbname = "nycflights"
+  )
+  on.exit(dbDisconnect(con))
+  
+  query <- glue::glue(
+    "SELECT {group_col}, AVG(arr_delay) as mean_delay from nycflights13.flights ",
+    "WHERE arr_delay > 0 GROUP BY {group_col}"
+  )
+  
+  return(dbGetQuery(con, query))
+}
+```
+
+## RSQLite
+```{r mean_delays_rsqlite, eval = FALSE}
+library(DBI)
+
+mean_delays <- function(group_col) {
+  con <- dbConnect(RSQLite::SQLite())
+  on.exit(dbDisconnect(con))
+  
+  query <- glue::glue(
+    "SELECT {group_col}, AVG(arr_delay) as mean_delay from nycflights13.flights ",
+    "WHERE arr_delay > 0 GROUP BY {group_col}"
+  )
+  
+  return(dbGetQuery(con, query))
+}
+```
+
+<div />
 
 If we give it the column `"month"`, we get the following dataframe:
 
@@ -94,7 +132,7 @@ test_that("mean_delays()", {
 
 And this works just fine if we only ever run your tests locally, but if we want to [run our tests with a Continuous Integration system](http://r-pkgs.had.co.nz/check.html#check) (and yes, we want to do that!), this won't work without first setting up our production database of flights. For our tests, we don't actually need to connect to the database and get new data (and, in fact, that would make some tests fail erroneously suddenly if the underlying changed). Instead, what we want is to take a snapshot of what happens when running the test code, and then be able to use that snapshot when we run tests later. These snapshots are frequently called fixtures (though you might hear people use other names like stubs or mocks).
 
-## Recording fixtures
[…]

[…]
