+++ tests/testthat/test-a-recording.R
@@ -7,7 +7,7 @@ test_that("We can capture db interactions", {
   expect_null(.dittodb_env$db_path)
 
   start_db_capturing()
-  con <- nycflights13_sqlite()
+  con <- nycflights13_create_sqlite()
 
   # our state has been updated
   expect_identical(.dittodb_env$db_path, file.path(temp_dir, "_memory_"))
@@ -36,7 +36,7 @@ test_that("We can capture db interactions", {
 test_that("We can specify the path when starting capture", {
   new_path <- file.path(temp_dir, "start_capture_path")
   start_db_capturing(new_path)
-  con <- nycflights13_sqlite()
+  con <- nycflights13_create_sqlite()
 
   # our state has been updated
   expect_identical(.dittodb_env$db_path, file.path(new_path, "_memory_"))
