+++ man/nycflights13_create_sql.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/nycflights13_sql.R
-\name{nycflights13_sql}
-\alias{nycflights13_sql}
+% Please edit documentation in R/nycflights13-sql.R
+\name{nycflights13_create_sql}
+\alias{nycflights13_create_sql}
 \title{Create a standardised database for testing}
 \usage{
-nycflights13_sql(con, schema = "", ...)
+nycflights13_create_sql(con, schema = "", ...)
 }
 \arguments{
 \item{con}{an SQL connection (i.e a PostgreSQL connection)}
@@ -51,10 +51,10 @@ connections <- list(
  )
 )
 
-nycflights13_sql(connections$odbc, schema = "nycflights13")
+nycflights13_create_sql(connections$odbc, schema = "nycflights13")
 # same as
-# nycflights13_sql(connections$rpostgresql, schema = "nycflights13")
+# nycflights13_create_sql(connections$rpostgresql, schema = "nycflights13")
 # also same as
-# nycflights13_sql(connections$rpostgres, schema = "nycflights13")
+# nycflights13_create_sql(connections$rpostgres, schema = "nycflights13")
 }
 }
