+++ man/nycflights_sqlite.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nycflights13-sql.R
+\name{nycflights_sqlite}
+\alias{nycflights_sqlite}
+\title{An SQLite connection to a subset of nycflights13}
+\usage{
+nycflights_sqlite()
+}
+\value{
+an RSQLiteConnection
+}
+\description{
+Included with \code{dittodb} is a small subset of \url{nycflights13}
+prepopulated into a \code{sqlite} database.
+}
+\details{
+This database is helpful for getting to know \code{dittodb} and running example
+code. It contains a small subset of the data in nycflights13: namely only the
+flights and planes that had a destination of ORD or MDW in February of 2013.
+The airports table has also been limited to only the New York and Chicago
+area airports.
+}
+\examples{
+con <- nycflights13_create_sqlite()
+}
