### 2020-06-06 1a48685
Some doc tweaks and adding examples
