+++ .github/workflows/check-mac.yaml
@@ -26,7 +26,7 @@ jobs:
     steps:
       - uses: actions/checkout@v1
 
-      - uses: r-lib/actions/setup-r@master
+      - uses: jonkeane/actions/setup-r@vers_failures
         with:
           r-version: ${{ matrix.config.r }}
 
