+++ .github/workflows/check-mac.yaml
@@ -14,7 +14,6 @@ jobs:
         config:
         - { os: macOS-latest, r: '3.6', args: "--no-manual" }
         - { os: macOS-latest, r: '4.0' }
-        - { os: macOS-latest, r: 'devel', args: "--no-manual" }
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
