### 2020-06-06 96457e4
Return to standard CI
