### 2020-06-07 e9acd5d
Add post-hoc functionality to remove empty TOC items that are generated by tabs generation
