### 2020-06-07 88c1efd
Add explicit 4.0.0 to pkgdown so it works without 4.0.1
