### 2020-06-07 6f8ce24
Remove a trailing whitespace
