### 2020-06-07 e09b37d
Documentation cleanup (#110)
* Tabbing in pkgdown!

* Add multiple db backends to readme/vignettes

* Some doc tweaks and adding examples

* Add post-hoc functionality to remove empty TOC items that are generated by tabs generation
