+++ man/nycflights13_create_sqlite.Rd
@@ -1,16 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/nycflights13_sql.R
-\name{nycflights13_sqlite}
-\alias{nycflights13_sqlite}
+% Please edit documentation in R/nycflights13-sql.R
+\name{nycflights13_create_sqlite}
+\alias{nycflights13_create_sqlite}
 \title{Create an in-memory SQLite database for testing}
 \usage{
-nycflights13_sqlite(location = ":memory:", ...)
+nycflights13_create_sqlite(location = ":memory:", ...)
 }
 \arguments{
 \item{location}{where to store the database}
 
 \item{...}{additional parameters to connect to a database (most are passed on
-to \code{\link{nycflights13_sql}})}
+to \code{\link{nycflights13_create_sql}})}
 }
 \value{
 RSQLiteConnection
@@ -20,6 +20,6 @@ Create an in-memory SQLite database for testing
 }
 \examples{
 \dontrun{
-con <- nycflights13_sqlite()
+con <- nycflights13_create_sqlite()
 }
 }

[…]
