+++ .github/workflows/pkgdown.yaml
@@ -27,4 +27,4 @@ jobs:
         run: R CMD INSTALL .
       - name: Deploy package
         run: |
-          Rscript -e "pkgdown:::deploy_local(new_process = FALSE, remote_url = 'https://x-access-token:${{secrets.GITHUB_TOKEN}}@github.com/${{github.repository}}.git')"
+          Rscript -e "pkgdown::deploy_to_branch(new_process = FALSE)"
