+++ pkgdown.yml
@@ -1,11 +1,12 @@
 pandoc: 2.7.3
-pkgdown: 1.4.1.9000
-pkgdown_sha: 732fb5b9ccc0ef87a73dfd0772ae661a4039b2a8
+pkgdown: 1.5.1.9000
+pkgdown_sha: d1a52e71e57dc8642fd0ae59a325cb9b6751f2bd
 articles:
   developing-dittodb: developing-dittodb.html
   dittodb: dittodb.html
   nycflights: nycflights.html
   travelling: travelling.html
+last_built: 2020-06-07T18:22Z
 urls:
   reference: https://jonkeane.com/dittodb/reference
   article: https://jonkeane.com/dittodb/articles
