### 2020-06-07 a8835ed
rename pkgdown to codemeta
