+++ .github/workflows/pkgdown-codemeta.yaml
@@ -1,6 +1,6 @@
 on:
   push:
-    branches: codemetadata
+    branches: master
 
 name: pkgdown+codemeta
 
