### 2020-06-07 4975065
Generate codemetadata on push to master (#111)
* Rebuilding codemeta along with pkgdown
