+++ .github/workflows/pkgdown-codemeta.yaml
@@ -2,10 +2,10 @@ on:
   push:
     branches: master
 
-name: pkgdown
+name: pkgdown+codemeta
 
 jobs:
-  pkgdown:
+  pkgdown-codemeta:
     runs-on: macOS-latest
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
@@ -22,11 +22,20 @@ jobs:
       - name: Install dependencies
         run: |
           Rscript -e 'install.packages("remotes")' \
-                  -e 'install.packages("ps")' \
+                  -e 'install.packages(c("ps", "codemetar"))' \
                   -e 'remotes::install_deps(dependencies = TRUE)' \
                   -e 'remotes::install_github("jonkeane/pkgdown@header_attributes")'
       - name: Install package
         run: R CMD INSTALL .
-      - name: Deploy package
+
+      - name: Render codemeta
+        run: Rscript -e 'codemetar::write_codemeta()'
+      - name: Commit results
+        run: |
+          git commit codemeta.json -m 'Re-build codemeta.json' || echo "No changes to commit"
+          git push https://${{github.actor}}:${{secrets.GITHUB_TOKEN}}@github.com/${{github.repository}}.git HEAD:${{ github.ref }} || echo "No changes to commit"
+          git pull
+
+      - name: Deploy pkgdown site
         run: |
           Rscript -e "pkgdown::deploy_to_branch(new_process = FALSE)"
