### 2020-06-12 2ad4d60
add dbExistsTable
