+++ man/mock-db-methods.Rd
@@ -1,11 +1,12 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/connection.R, R/dbListTables-Fields.R,
-%   R/dbQueries-Results.R, R/dbMisc.R, R/quote.R
+% Please edit documentation in R/connection.R, R/dbExistsTable.R,
+%   R/dbListTables-Fields.R, R/dbQueries-Results.R, R/dbMisc.R, R/quote.R
 \docType{class}
 \name{mock-db-methods}
 \alias{mock-db-methods}
 \alias{DBIMockConnection-class}
 \alias{dbDisconnect,DBIMockConnection-method}
+\alias{dbExistsTable,DBIMockConnection,character-method}
 \alias{dbListTables,DBIMockConnection-method}
 \alias{dbListFields,DBIMockConnection,character-method}
 \alias{dbListFields,DBIMockConnection,Id-method}
@@ -29,6 +30,8 @@
 \title{Methods for interacting with DB mocks instead of an actual database}
 \usage{
 \S4method{dbDisconnect}{DBIMockConnection}(conn, ...)
+
+\S4method{dbExistsTable}{DBIMockConnection,character}(conn, name, ...)
 
 \S4method{dbListTables}{DBIMockConnection}(conn, ...)
 
