+++ NAMESPACE
@@ -33,6 +33,7 @@ exportClasses(DBIMockSQLiteConnection)
 exportMethods(dbClearResult)
 exportMethods(dbColumnInfo)
 exportMethods(dbDisconnect)
+exportMethods(dbExistsTable)
 exportMethods(dbFetch)
 exportMethods(dbGetInfo)
 exportMethods(dbGetQuery)
