+++ tests/testthat/test-dbi-generic-integration.R
@@ -159,8 +159,32 @@ for (pkg in names(db_pkgs)) {
       dbClearResult(result)
 
       # dbExistsTable ====
-      table_exists <- dbExistsTable(con, "airlines")
-      table_does_not_exist <- dbExistsTable(con, "doesnotexist")
+      # dbExistsTable is ever so slightly different for each
+      if (pkg == "RMariaDB") {
+        table_exists <- dbExistsTable(con, "airlines")
+        table_does_not_exist <- dbExistsTable(con, "doesnotexist")
+      } else if (pkg == "odbc") {
+        table_exists <- dbExistsTable(
+          con,
+          Id(schema = schema, table = "airlines")
+        )
+        table_does_not_exist <- dbExistsTable(
+          con,
+          Id(schema = schema, table = "doesnotexist")
+        )
+      } else if (pkg == "RPostgreSQL") {
+        table_exists <- dbExistsTable(con, c(schema, "airlines"))
+        table_does_not_exist <- dbExistsTable(con, c(schema, "doesnotexist"))
+      } else if (pkg == "RPostgres") {
+        table_exists <- dbExistsTable(
+          con,
+          Id(schema = schema, table = "airlines")
+        )
+        table_does_not_exist <- dbExistsTable(
+          con,
+          Id(schema = schema, table = "doesnotexist")
+        )
+      }
 
       dbDisconnect(con)
       stop_db_capturing()
@@ -306,8 +330,31 @@ for (pkg in names(db_pkgs)) {
 
         # dbExistsTable ====
         test_that("dbGetInfo", {
-          expect_true(dbExistsTable(con, "airlines"))
-          expect_false(dbExistsTable(con, "doesnotexist"))
+          if (pkg == "RMariaDB") {
+            expect_true(table_exists <- dbExistsTable(con, "airlines"))
+            expect_false(table_does_not_exist <- dbExistsTable(con, "doesnotexist"))
+          } else if (pkg == "odbc") {
+            expect_true(table_exists <- dbExistsTable(
+              con,
+              Id(schema = schema, table = "airlines")
+            ))
+            expect_false(table_does_not_exist <- dbExistsTable(
+              con,
+              Id(schema = schema, table = "doesnotexist")
+            ))
+          } else if (pkg == "RPostgreSQL") {
+            expect_true(table_exists <- dbExistsTable(con, c(schema, "airlines")))
+            expect_false(table_does_not_exist <- dbExistsTable(con, c(schema, "doesnotexist")))
+          } else if (pkg == "RPostgres") {
+            expect_true(table_exists <- dbExistsTable(
+              con,
+              Id(schema = schema, table = "airlines")
+            ))
+            expect_false(table_does_not_exist <- dbExistsTable(
+              con,
+              Id(schema = schema, table = "doesnotexist")
+            ))
+          }
         })
 
         # dbWriteTable ====
