### 2020-06-12 60eb8b1
Need to unmock `dbExistsTable` too
