+++ .github/workflows/pkgdown-codemeta.yaml
@@ -13,7 +13,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@master
         with:
-          r-version: '4.0.0'
+          r-version: '4.0'
       - uses: r-lib/actions/setup-pandoc@master
       - name: Install databases (macOS)
         run: |
