+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   dittodb: dittodb.html
   nycflights: nycflights.html
   travelling: travelling.html
-last_built: 2020-06-12T16:55Z
+last_built: 2020-06-12T17:13Z
 urls:
   reference: https://jonkeane.com/dittodb/reference
   article: https://jonkeane.com/dittodb/articles
