+++ R/dbExistsTable.R
@@ -1,6 +1,6 @@
 .dbExistsTable <- function(conn, name, ...) {
   name <- sanitize_table_id(name, ...)
-  return(list_helper(glue("dbExistsTable-{name}"), conn, ...))
+  return(list_helper(glue::glue("dbExistsTable-{name}"), conn, ...))
 }
 
 #' @rdname mock-db-methods
