### 2020-06-16 b745142
Update install directions
