### 2020-06-16 735897c
Add branch option for codemetar
