+++ .github/workflows/check-windows.yaml
@@ -17,7 +17,7 @@ jobs:
     env:
       CRAN: ${{ matrix.config.cran }}
       DBROOTPASS: r2N5y7V*
-      DITTODB_ENABLE_PG_TESTS: true
+      DITTODB_ENABLE_PG_TESTS: false
       DITTODB_ENABLE_MARIA_TESTS: true
 
     steps:
@@ -50,7 +50,6 @@ jobs:
         run: |
           $startingpath = [System.Environment]::GetEnvironmentVariable("Path","Machine")
           choco install mariadb
-          choco uninstall postgresql
           choco install postgresql --force --params '/Password:${{ env.DBROOTPASS }}' --paramsglobal
           choco install psqlodbc
 
