### 2020-06-16 2a593fb
Update check-windows.yaml
