+++ .github/workflows/check-windows.yaml
@@ -1,4 +1,4 @@
-on: [pull_request]
+on: [push, pull_request]
 
 name: check-windows
 
@@ -50,7 +50,7 @@ jobs:
         run: |
           $startingpath = [System.Environment]::GetEnvironmentVariable("Path","Machine")
           choco install mariadb
-          choco install postgresql10 --params '/Password:${{ env.DBROOTPASS }}' --paramsglobal
+          choco install postgresql11 --params '/Password:${{ env.DBROOTPASS }}' --paramsglobal
           choco install psqlodbc
 
           # add choco-added paths to the actions path
