### 2020-06-23 #366 @jonkeane
Hello everyone,

I hope everyone is doing ok. I very much know that everyone is busy, but I was hoping to start the submission process for `dittodb` to CRAN in the near future. I know that the reviews are separate from CRAN submission, but wanted to make sure there wasn't anything major or glaring from the reviewers before starting that process. Thoughts?
