### 2020-06-25 #366 @etiennebr
I don't want to hold back the process, and I'm not sure if @annakrystalli is waiting for additional comments from us?

I think the package should be released to CRAN and join rOpenSci. I haven't re-tried the package, but I've looked at many of the changes and the responses from @jonkeane and I think the package should go reach its audience and live. Look at the new fancy [README,](https://dittodb.jonkeane.com/) with thumbnails to select for the database driver! I probably won't have the time to try the package before at least mid-July, so is there something I can do meanwhile?
