+++ NEWS.md
@@ -1,4 +1,4 @@
-# dittodb 0.0.1 (under development)
+# dittodb 0.0.1 
 * Initial release with functionality for recording and playing back database fixtures from a number of DBI-based drivers (RSQLite, RPostgres, RMariaDB, RPostgreSQL)
 * Thanks to @maelle for PR#12
 * `nycflights13_create_sql()` now always uses DBI
