+++ .github/workflows/pkgdown-codemeta.yaml
@@ -22,7 +22,8 @@ jobs:
       - name: Install dependencies
         run: |
           Rscript -e 'install.packages("remotes")' \
-                  -e 'install.packages(c("ps", "codemetar"))' \
+                  -e 'install.packages(c("ps")' \
+                  -e 'remotes::install_github("ropensci/codemetar")'
                   -e 'remotes::install_deps(dependencies = TRUE)' \
                   -e 'remotes::install_github("jonkeane/pkgdown@header_attributes")'
       - name: Install package
