+++ cran-comments.md
@@ -1,0 +1,17 @@
+## Resubmission
+* I've corrected 3 (possibly) invalid file URIs in 
+  man/nycflights13_create_sql.Rd, man/nycflights_sqlite.Rd, and README.md
+
+
+## Test environments
+* local R installation, R 4.0.0
+* ubuntu 18.04 (on GitHub actions), R 3.3-4.0
+* macOS (on GitHub actions), R 3.6, 4.0, devel
+* windows (on GitHub actions), R 4.0
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* The note is that this is a new release.
+
