### 2020-07-14 #366 @etiennebr
I'm happy! I recommend this package to rOpenSci. Good job @jonkeane!
