+++ db-setup/mariadb-brew.sh
@@ -2,6 +2,6 @@ rm -f /etc/my.cnf
 brew install mariadb
 brew services start mariadb
 mkdir /usr/local/etc/my.cnf.d
-sleep 2
+sleep 10
 # TODO: add dbrootpass if it doesn't exist?
 sudo mysql -u root -e "ALTER USER 'root'@'localhost' IDENTIFIED BY '$DBROOTPASS';"
