### 2020-07-15 #366 @annakrystalli
Approved! 🚀

Thanks @jonkeane for submitting and @etiennebr & @helenmiller16  for your reviews! 😸

To-dos:
- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I will shortly invite you to a team that should allow you to do so.  You'll be made admin once you do.
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
