### 2020-07-18 a9c359d
all of ropensci requested changes except pkgdown tweaks deactivation
