+++ README.md
@@ -1,9 +1,9 @@
 # dittodb
 <!-- badges: start -->
-[![Windows](https://github.com/jonkeane/dittodb/workflows/check-windows/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-windows)
-[![Mac](https://github.com/jonkeane/dittodb/workflows/check-mac/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-mac)
-[![Linux](https://github.com/jonkeane/dittodb/workflows/check-linux-ubuntu/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-linux-ubuntu)
-[![Codecov test coverage](https://codecov.io/gh/jonkeane/dittodb/branch/main/graph/badge.svg)](https://codecov.io/gh/jonkeane/dittodb?branch=main)
+[![Windows](https://github.com/ropensci/dittodb/workflows/check-windows/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-windows)
+[![Mac](https://github.com/ropensci/dittodb/workflows/check-mac/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-mac)
+[![Linux](https://github.com/ropensci/dittodb/workflows/check-linux-ubuntu/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-linux-ubuntu)
+[![Codecov test coverage](https://codecov.io/gh/ropensci/dittodb/branch/main/graph/badge.svg)](https://codecov.io/gh/ropensci/dittodb?branch=main)
 [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](https://www.r-pkg.org/badges/version/dittodb)](https://CRAN.R-project.org/package=dittodb)
 <!-- badges: end -->
@@ -133,7 +133,7 @@ All without having to ever set a database up on Travis 🎉
 Currently, dittodb is not on CRAN. You can install from source, or use `remotes`:
 
 ```r
-remotes::install_github("jonkeane/dittodb@main")
+remotes::install_github("ropensci/dittodb@main")
 ```
 
 _Note_ the `@main` at the end is necessary until the `remotes` package 
@@ -144,7 +144,7 @@ currently uses.
 An alternative method of installation is to use `remotes::git()` directly:
 
 ```r
-remotes::install_git("https://github.com/jonkeane/dittodb.git")
+remotes::install_git("https://github.com/ropensci/dittodb.git")
 ```
 
 ## Setup a package to use `dittodb`
@@ -154,7 +154,7 @@ to `tests/testthat/helper.R`.
 
 ## Development
 
-There is extensive information about developing `dittodb` in the vignette [Developing `dittodb`](https://dittodb.jonkeane.com/articles/developing-dittodb.html, please read that before trying to make changes to `dittodb` or running any of the scripts provided in the `db-setup` directory.
+There is extensive information about developing `dittodb` in the vignette [Developing `dittodb`](https://docs.ropensci.org/dittodb/articles/developing-dittodb.html, please read that before trying to make changes to `dittodb` or running any of the scripts provided in the `db-setup` directory.
 
 In order to test `dittodb` recording functionality locally or on CI, it is helpful to have databases with test data available. This can be accomplished using the scripts in the `db-setup` directory. By default, `dittodb` does not run any tests that require database infrastructure locally.
 
@@ -164,4 +164,4 @@ On continuous integration, (using GitHub Actions) these scripts in the `db-setup
 
 ## Code of Conduct
 
[…]
