### 2020-07-18 #366 @jonkeane
Thanks for this, and sorry for my delay in getting back. I will move my repo under the organization shortly. Do you know if there will be a delay in being able to interact with the repo after I transfer it but before I'm an admin? If so, would it be possible to coordinate a time to do that? I'm doing some work on it this weekend and don't want to accidentally lock myself out.

As for the docs, I think at least now I would like to stick to hosting them myself — there's a bit fo customization I had to do to the template to get it working.

I'll work on a blog post as well. I've already got a few examples ready and part of a blog post that I'm working on for once {dittodb} is accepted to CRAN.

Thank you to everyone, @annakrystalli for the hard work of shepherding us all, @helenmiller16 and @etiennebr for really great comments that improved {dittodb}, and @pachamaltese for help, support, and code
