### 2020-07-18 5cca904
First pass of changes
