+++ R/connection.R
@@ -1,9 +1,10 @@
 #' Methods for interacting with DB mocks instead of an actual database
 #'
-#' Various methods (`dbSendQuery`, `dbFetchQuery`) that are mocks of the DBI
-#' methods of the same name. Instead of actually interacting with a database,
-#' they read in mock responses and the code proceeds after that. These aren't
-#' used directly, but are part of how `dittodb` works.
+#' Various methods (`dbSendQuery`, `dbFetchQuery`) that are mocks of the
+#' [DBI](https://CRAN.R-project.org/package=DBI) methods of the same name.
+#' Instead of actually interacting with a database, they read in mock responses
+#' and the code proceeds after that. These aren't used directly, but are part of
+#' how {dittodb} works.
 #'
 #' @param conn a database connection (for dispatch with these methods, it should
 #'   be of class `DBIMockConnection`)
