+++ R/nycflights13-sql.R
@@ -150,14 +150,15 @@ nycflights13_create_sqlite <- function(location = ":memory:", ...) {
 
 #' An SQLite connection to a subset of nycflights13
 #'
-#' Included with `dittodb` is a small subset of [`nycflights13`](https://CRAN.R-project.org/package=nycflights13)
+#' Included with {dittodb} is a small subset of
+#' [`nycflights13`](https://CRAN.R-project.org/package=nycflights13)
 #' prepopulated into a `sqlite` database.
 #'
-#' This database is helpful for getting to know `dittodb` and running example
+#' This database is helpful for getting to know {dittodb} and running example
 #' code. It contains a small subset of the data in nycflights13: namely only the
-#' flights and planes that had a destination of ORD or MDW in February of 2013.
-#' The airports table has also been limited to only the New York and Chicago
-#' area airports.
+#' flights and planes that had a destination of ORD or MDW (the codes for the
+#' two major airports in Chicago) in February of 2013. The airports table has
+#' also been limited to only the New York and Chicago area airports.
 #'
 #' @return an RSQLiteConnection
 #'
