+++ man/mockdb.Rd
@@ -24,7 +24,7 @@ Wrap a chunk of code in \code{with_mock_db()} to use mocked databases that will
 use fixtures instead of connecting to a real database. Alternatively, you can
 start and stop using a mocked database with \code{start_mock_db()} and
 \code{stop_mock_db()} respectively.to execute the whole thing without needing to
-remember to stop the mocking. When testing with \code{dittodb}, it will look for
+remember to stop the mocking. When testing with {dittodb}, it will look for
 fixtures in all entries of \code{\link{db_mock_paths}}.
 }
 \details{
@@ -38,7 +38,7 @@ step through tests interactively using \code{start_mock_db()}/\code{stop_mock_db
 
 Connections should be made after \code{start_mock_db()} if you're using that
 function or they should be made inside of \code{with_mock_db()} if you're using
-that function because \code{dittodb} uses the database name (given in \code{dbname} or
+that function because {dittodb} uses the database name (given in \code{dbname} or
 \code{Database} argument of \code{\link{dbConnect}} depending on the driver) to separate
 different fixtures.
 }
