### 2020-07-18 #366 @pachamaltese
@annakrystalli Hi!! Is any chance to use custom pkgdown settings?
