+++ README.md
@@ -154,7 +154,7 @@ to `tests/testthat/helper.R`.
 
 ## Development
 
-There is extensive information about developing `dittodb` in the vignette [Developing `dittodb`](https://docs.ropensci.org/dittodb/articles/developing-dittodb.html, please read that before trying to make changes to `dittodb` or running any of the scripts provided in the `db-setup` directory.
+There is extensive information about developing `dittodb` in the vignette [Developing `dittodb`](https://dittodb.jonkeane.com//articles/developing-dittodb.html, please read that before trying to make changes to `dittodb` or running any of the scripts provided in the `db-setup` directory.
 
 In order to test `dittodb` recording functionality locally or on CI, it is helpful to have databases with test data available. This can be accomplished using the scripts in the `db-setup` directory. By default, `dittodb` does not run any tests that require database infrastructure locally.
 
@@ -164,4 +164,4 @@ On continuous integration, (using GitHub Actions) these scripts in the `db-setup
 
 ## Code of Conduct
 
-Please note that the `dittodb` project is released with a [Contributor Code of Conduct](https://docs.ropensci.org/dittodb/CODE_OF_CONDUCT). By contributing to this project, you agree to abide by its terms.
+Please note that the `dittodb` project is released with a [Contributor Code of Conduct](https://dittodb.jonkeane.com//CODE_OF_CONDUCT). By contributing to this project, you agree to abide by its terms.
