### 2020-07-18 #366 @pachamaltese
Thanks a lot @jonkeane, I am really glad to be the Sideshow Pacha on this project. I've sent a PR that implements all of the requested changes except for the pkgdown parts. IMO, the tabs really add a lot of value and make the documentation a lot easier to read.
