### 2020-07-18 #366 @jonkeane
Of course, and thanks for all of your help. I don't think there's any need to complicate the ROpenSci docs process just for {dittodb} regardless, I think at least for now I would like to keep them where they are regardless (dittodb.jonkeane.com) since I've used that URL in advertising, etc.
