### 2020-07-18 d0547c5
More CRAN revisions
