+++ R/mock-db.R
@@ -30,7 +30,6 @@
 #' @name mockdb
 #'
 #' @examples
-#' \dontrun{
 #' # Add the mocks included with dittodb to the db_mock_paths to use them below
 #' db_mock_paths(system.file("nycflight_mocks", package = "dittodb"), last = TRUE)
 #'
@@ -41,42 +40,40 @@
 #'     dbname = "nycflights"
 #'   )
 #'
-#'   test_that("We get one airline", {
+#'   testthat::test_that("We get one airline", {
 #'     one_airline <- dbGetQuery(
 #'       con,
 #'       "SELECT carrier, name FROM airlines LIMIT 1"
 #'     )
-#'     expect_is(one_airline, "data.frame")
-#'     expect_equal(nrow(one_airline), 1)
-#'     expect_equal(one_airline$carrier, "9E")
-#'     expect_equal(one_airline$name, "Endeavor Air Inc.")
+#'     testthat::expect_is(one_airline, "data.frame")
+#'     testthat::expect_equal(nrow(one_airline), 1)
+#'     testthat::expect_equal(one_airline$carrier, "9E")
+#'     testthat::expect_equal(one_airline$name, "Endeavor Air Inc.")
 #'   })
+#'
+#'   dbDisconnect(con)
 #' })
 #'
-#' # using `start_mock_db()` and `stop_mock_db()` (and a different driver)
+#' # using `start_mock_db()` and `stop_mock_db()`
 #' start_mock_db()
-#' con <- dbConnect(
-#'   RPostgreSQL::PostgreSQL(),
-#'   dbname = "nycflights",
-#'   host = "127.0.0.1",
-#'   user = "some_user",
-#'   password = ""
-#' )
+#'   con <- dbConnect(
+#'     RSQLite::SQLite(),
+#'     dbname = "nycflights"
+#'   )
 #'
-#' test_that("We get one airline", {
+#' testthat::test_that("We get one airline", {
 #'   one_airline <- dbGetQuery(
 #'     con,
 #'     "SELECT carrier, name FROM airlines LIMIT 1"
 #'   )
-#'   expect_is(one_airline, "data.frame")
-#'   expect_equal(nrow(one_airline), 1)
-#'   expect_equal(one_airline$carrier, "9E")
-#'   expect_equal(one_airline$name, "Endeavor Air Inc.")
+#'   testthat::expect_is(one_airline, "data.frame")
+#'   testthat::expect_equal(nrow(one_airline), 1)
+#'   testthat::expect_equal(one_airline$carrier, "9E")
+#'   testthat::expect_equal(one_airline$name, "Endeavor Air Inc.")
 #' })
 #'
 #' dbDisconnect(con)
 #' stop_mock_db()
-#' }
 with_mock_db <- function(expr) {
   # TODO: change this to use start_mock_db()?
   with_mock(
