+++ man/capture_requests.Rd
@@ -52,6 +52,7 @@ block. When you connect to the database, dittodb sets up the mocks for the
 specific database you're connecting to when you call \code{\link[DBI:dbConnect]{DBI::dbConnect}}.
 }
 \examples{
+\donttest{
 # Temporary files for examples
 nycflights_path <- tempfile()
 
@@ -86,4 +87,5 @@ with_mock_db({
   # the result from df3 above
   print(dbGetQuery(con, "SELECT * FROM airlines LIMIT 3"))
 })
+}
 }
