### 2020-07-19 fcd0344
Allow notes on R 3.3 (because it didn't (yet) know about the rev role
