+++ .github/workflows/check-linux-ubuntu.yaml
@@ -84,7 +84,10 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '${{ matrix.config.args }}', error_on = 'note', check_dir = 'check')"
+        run: |
+          # Only error on warnings or above for anything prior to 3.4
+          errornon=$(if (( $(echo "${{ matrix.config.args.r }} < 3.4" | bc -l) )); then echo "warning"; else echo "note"; fi)
+          Rscript -e "rcmdcheck::rcmdcheck(args = '${{ matrix.config.args }}', error_on = 'errornon', check_dir = 'check')"
 
       - name: Upload check results
         if: failure()
