### 2020-07-19 4b5c746
Small typo in vignette
