### 2020-07-19 aa80c2a
CRAN comments (#123)
