+++ tests/testthat/test-mock-paths.R
@@ -28,6 +28,15 @@ withr::with_options(
     test_that("dittodb.mock.paths is used", {
       expect_identical(db_mock_paths(), "new_one")
     })
+  })
+
+with_mock_path(
+  c("new_one"),
+  replace = TRUE,
+  {
+    test_that("the one mockpath is used", {
+      expect_identical(db_mock_paths(), "new_one")
+    })
 
     test_that("and we can unset", {
       db_mock_paths(NULL)
@@ -44,6 +53,13 @@ withr::with_options(
     test_that("and we can add to the end", {
       db_mock_paths("end", last = TRUE)
       expect_identical(db_mock_paths(), c("tests/testthat/", ".", "end"))
+      # cleanup
+      db_mock_paths(NULL)
+    })
+
+    test_that("and we can replace", {
+      db_mock_paths("replace", replace = TRUE)
+      expect_identical(db_mock_paths(), c("replace"))
       # cleanup
       db_mock_paths(NULL)
     })
