### 2020-07-19 865c5f4
More speed with packagemanager
