+++ .github/workflows/check-linux-ubuntu.yaml
@@ -12,11 +12,11 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: ubuntu-18.04, r: '3.3', args: "--no-manual" }
-        - { os: ubuntu-18.04, r: '3.4', args: "--no-manual" }
-        - { os: ubuntu-18.04, r: '3.5', args: "--no-manual" }
-        - { os: ubuntu-18.04, r: '3.6', args: "--no-manual" }
-        - { os: ubuntu-18.04, r: '4.0', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '3.3', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
+        - { os: ubuntu-18.04, r: '3.4', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
+        - { os: ubuntu-18.04, r: '3.5', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
+        - { os: ubuntu-18.04, r: '3.6', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
+        - { os: ubuntu-18.04, r: '4.0', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
 
     env:
       CRAN: ${{ matrix.config.cran }}
