### 2020-07-19 cb4122b
Use line wrapping in readme
