+++ .github/workflows/check-linux-ubuntu.yaml
@@ -12,17 +12,18 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: ubuntu-18.04, r: '3.3', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
-        - { os: ubuntu-18.04, r: '3.4', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
-        - { os: ubuntu-18.04, r: '3.5', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
-        - { os: ubuntu-18.04, r: '3.6', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
-        - { os: ubuntu-18.04, r: '4.0', args: "--no-manual", rspm: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest" }
+        - { os: ubuntu-18.04, r: '3.3', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '3.4', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '3.5', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '3.6', args: "--no-manual" }
+        - { os: ubuntu-18.04, r: '4.0', args: "--no-manual" }
 
     env:
       CRAN: ${{ matrix.config.cran }}
       DBROOTPASS: r2N5y7V*
       DITTODB_ENABLE_PG_TESTS: true
       DITTODB_ENABLE_MARIA_TESTS: true
+      RSPM: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest"
 
     steps:
       - uses: actions/checkout@v1
