### 2020-07-19 7cfdacf
all of ropensci requested changes except pkgdown tweaks deactivation (#122)
