+++ man/mockPaths.Rd
@@ -32,7 +32,7 @@ working directory (or the test directory). If you want to look in other
 places, you can call \code{db_mock_paths} to add directories to the search path.
 }
 \details{
-It works like \code{\link[base:libPaths]{base::.libPaths()}}: any directories you specify will be added
+It works like \code{\link[base:.libPaths]{base::.libPaths()}}: any directories you specify will be added
 to the list and searched first. The default directory will be searched last.
 Only unique values are kept: if you provide a path that is already found in
 \code{db_mock_paths}, the result effectively moves that path to the first
