+++ README.md
@@ -1,9 +1,9 @@
 # dittodb
 <!-- badges: start -->
-[![Windows](https://github.com/jonkeane/dittodb/workflows/check-windows/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-windows)
-[![Mac](https://github.com/jonkeane/dittodb/workflows/check-mac/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-mac)
-[![Linux](https://github.com/jonkeane/dittodb/workflows/check-linux-ubuntu/badge.svg)](https://github.com/jonkeane/dittodb/actions?workflow=check-linux-ubuntu)
-[![Codecov test coverage](https://codecov.io/gh/jonkeane/dittodb/branch/main/graph/badge.svg)](https://codecov.io/gh/jonkeane/dittodb?branch=main)
+[![Windows](https://github.com/ropensci/dittodb/workflows/check-windows/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-windows)
+[![Mac](https://github.com/ropensci/dittodb/workflows/check-mac/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-mac)
+[![Linux](https://github.com/ropensci/dittodb/workflows/check-linux-ubuntu/badge.svg)](https://github.com/ropensci/dittodb/actions?workflow=check-linux-ubuntu)
+[![Codecov test coverage](https://codecov.io/gh/ropensci/dittodb/branch/main/graph/badge.svg)](https://codecov.io/gh/ropensci/dittodb?branch=main)
 [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![CRAN status](https://www.r-pkg.org/badges/version/dittodb)](https://CRAN.R-project.org/package=dittodb)
 <!-- badges: end -->
@@ -130,52 +130,32 @@ with_mock_db({
 All without having to ever set a database up on Travis or GitHub Actions 🎉
 
 ## Installation
-Currently, {dittodb} is not on CRAN (The Comprehensive R Archive Network). You 
-can install from source, or use {remotes}:
+Currently, {dittodb} is not on CRAN (The Comprehensive R Archive Network). You can install from source, or use {remotes}:
 
 ```r
-remotes::install_github("jonkeane/dittodb@main")
+remotes::install_github("ropensci/dittodb@main")
 ```
 
-_Note_ the `@main` at the end is necessary until the {remotes} package 
-[is updated](https://github.com/r-lib/remotes/issues/508) to use the default 
-branch specified for GitHub repositories rather than the hard coded name it 
-currently uses.
+_Note_ the `@main` at the end is necessary until the {remotes} package [is updated](https://github.com/r-lib/remotes/issues/508) to use the default branch specified for GitHub repositories rather than the hard coded name it currently uses.
 
 An alternative method of installation is to use `remotes::git()` directly:
 
 ```r
-remotes::install_git("https://github.com/jonkeane/dittodb.git")
+remotes::install_git("https://github.com/ropensci/dittodb.git")
 ```
 
 ## Setup a package to use {dittodb}
-Use the function `dittodb::use_dittodb()` to easily get started using {dittodb}. 
-It will add {dittodb} to `Suggests` in the `DESCRIPTION` file and add `library(dittodb)`
-to `tests/testthat/helper.R`.
+Use the function `dittodb::use_dittodb()` to easily get started using {dittodb}. It will add {dittodb} to `Suggests` in the `DESCRIPTION` file and add `library(dittodb)` to `tests/testthat/helper.R`.
 
 ## Development
[…]
