+++ .github/workflows/check-linux-ubuntu.yaml
@@ -23,6 +23,7 @@ jobs:
       DBROOTPASS: r2N5y7V*
       DITTODB_ENABLE_PG_TESTS: true
       DITTODB_ENABLE_MARIA_TESTS: true
+      RSPM: "https://packagemanager.rstudio.com/all/__linux__/bionic/latest"
 
     steps:
       - uses: actions/checkout@v1
@@ -84,7 +85,10 @@ jobs:
           MYSQL_PWD: ${{ env.DBROOTPASS }}
 
       - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = '${{ matrix.config.args }}', error_on = 'note', check_dir = 'check')"
+        run: |
+          # Only error on warnings or above for anything prior to 3.4
+          erroron=$(if (( $(echo "${{ matrix.config.r }} < 3.4" | bc -l) )); then echo "warning"; else echo "note"; fi)
+          Rscript -e "rcmdcheck::rcmdcheck(args = '${{ matrix.config.args }}', error_on = '$erroron', check_dir = 'check')"
 
       - name: Upload check results
         if: failure()
