+++ reference/tests/testthat/_var_folders_24_8k48jl6d249_n_qfxwsl6xvm0000gn_T__RtmpVeOvO8_filee05600fdc0/SELECT-f2090b.R
@@ -1,0 +1,3 @@
+structure(list(carrier = c("9E", "AA"), name = c("Endeavor Air Inc.", 
+"American Airlines Inc.")), class = "data.frame", row.names = c(NA, 
+-2L))
