### 2020-07-19 #366 @jonkeane
Ok, after reading through the docs on transferring, I've gone ahead and transferred the repo to https://github.com/ropensci/dittodb
