### 2020-07-19 1a03707
readme update, docs
