### 2020-07-19 e8a9cb4
More cran request changes
