+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: dittodb
-Title: A Test Environment for database Requests
+Title: A Test Environment for Database Requests
 Version: 0.1.0
 Authors@R: c(
     person(given = "Jonathan",
