### 2020-07-19 66c5b5a
Update cran-comments
