### 2020-07-19 #366 @annakrystalli
Hello @jonkeane ! Thanks for transferring the repo. I've now made you admin again.

Re the docs, I'm just enquiring as to whether anything needs to be done so as NOT to use the rOpenSci docs. Will let you know when I know :)

In the meantime, any chance you could check off from the above checklist what has been completed already? 😎🙏
