### 2020-07-19 #366 @jonkeane
Thank you so much! I don't know if this is me being dense, but I don't seem to be able to tick those boxes myself (I've tried both the box itself and clicking the three dots more message above that...)

In any event, what's been done now are all of the stages except the pkgdown one:

* ☑️ Transfer the repo to...
* ☑️ Fix all links to the GitHub repo...
* ☑️ Fix any links in badges for CI...
* ☑️ We're starting to roll out software metadata...
