### 2020-07-20 #366 @annakrystalli
Oh! Sorry about that. Didn't realise you could tick the boxes! Good to know.

Regarding the docs, there is nothing else you need to do and it's fine to carry on hosting them. The rOpenSci docs will still be built but so long as you direct users to your own docs, that shouldn't be a problem.

So, I think that's wrap! I'll close this issue now.

Thanks again to everyone involved!
